Studio friendly support library
Projects with api specific source sets are a problem for Android
Studio because it cannot find them in sources list.

This CL adds the ability to convert these API specific source sets
into modules when the project is opened inside the IDE.
When run from command line, it still fallbacks to the previous
implementation.

For projects with resources, we still rely on make to do the
necessary checks.

Bug: 27567189
Change-Id: I1d0be60f1b2289a4b006c196b4992a2de3868336


GitOrigin-RevId: 3986e04c740dc8fe51e7892ddcba0c9a17dee117