Remove uses of @ParallelizableTask annotation.
ag/2428885 removed @ParallelizableTask annotation and so we no longer
compile with the latest plugin. This annotation was an incubating feature
and gradle 4.0 removed support for it as they moved to a new structure
for parallel tasks. These annotations already did nothing since our switch
to 4.0. Note, that these annotations were for parallel runs within the same
module, cross module parallelism continues to work.

Test: ./gradlew assembleDebug (using newest AGP)
Change-Id: Iecbad82947be17355fc4ec6b6d5880c4f802f726


GitOrigin-RevId: 864cb4a0cc946aafc5d73f21400cec301723f47b