Move to AGP 3.0.0 stable 😁
Finally after many months of pain, suffering and tears we get to
move to the stable version of AGP even it this is going to last
us a short while until we need the next bleeding edge AGP feature
in support library to make our lives gloomy yet again.

For now - rejoice!

Test: ./gradlew buildOnServer
Change-Id: Ia6e1f7a8ecc4ecc4e79283d970d6defbd70828c6


GitOrigin-RevId: 5cb9a17d2a5e8622b9a3737a278d7d96b7cc6865