enable yamlV2 output for support library, for compatibility with new DAC
This flag adds an extra file to the output: _NEW_book.yaml, which is
a _book.yaml file for the site that contains the complete list of API
links (instead of only the package names). This file will be ignored by
the current version of DAC, but for our upcoming redesign, this file
must be renamed _toc.yaml. More info in ag/3382461

bug: 62869488
Change-Id: I14a55ab583c955062512fe493a05d7a9144ba2ca


GitOrigin-RevId: 33e006558146e45581b762322ae3458486c8a5b1