Handle Kotlin to Java Relations
When a Kotlin POJO references a relation entity that was defined in
Java, it generates a wildcard type, which triggers a compilation
failure.

This CL fixes the issue by dropping the wildcard while processing.
We probably need to change other places where we might read wildcard
as well but we are at beta2 so it would be a risky release at this
point. + this only happens in a weird case where code is generated
by kotlin.

Also set the version to beta3 for the release and removed some
unnecessary schemas from the kotlin demo

Bug:78199923
Test: PojoProcessorTest, QueryMethodProcessorTest, BooksDao (for
codegen)
Change-Id: I91e4facdfaf0d0f0e3d56ccc443e1b1c1d423c94


GitOrigin-RevId: 5e4d488036d51131bf4c97102fed7905d88d086d