Use cannonical path while resolving folders in ProjectGraph
This CL changes ProjectGraph to use cannonical paths when resolving
project paths. Not doing so breaks the test on MAC where tmp folders
get a /private suffix only when used in the root project (seems like
root project is the only one using cannonical path). We now always
use the cannonical file to ensure that we resolve any symlinks.

Bug: 112205561
Test: ProjectGraphTest
Change-Id: I86cbd42bd06eaafa596a98769844bf1760106ab2


GitOrigin-RevId: 3ac0c8b8982756630476b0c2ab0ca07e13010391