Support for testing different dependency versions
AndroidX libraries should now declare their dependencies as a module
with a fixed version instead of a local project (e.g
"androidx.collection:collection:1.0.0" instead of ":collection".) The
version specified should be the minimum supported version the library
depends on.

If project A contains any (prebuilt) dependencies of the form
androidx.B:B:version, then A will be built twice: once depending on
(the prebuilt) androidx.B:B:version , and once depending on
(the project) ":B", in an effort to check the compatibility expected
by the SemVer policy for B.

Test: manual testing.
Test: testing to ensure that we only substitute fixed version with
current projects that actually exist and are within the same major
release.

Bug: 112531649
Change-Id: Ic3539e2bbf552ebcc4559c34f8724e9659c968f1


GitOrigin-RevId: 75a03ec0bfe2b70a2a449e22f217daa74cfb1adc