Added support for custom error prone rules
We are now able to write our own custom error prone rules as java
files inside the customerrorprone project. There are currently
3 custom rules:
- BanUnlistedSDKVersionComparison bans comprison of SDK_INT
to anything other than constants in VERSION_CODES. This
rule is written but not yet enforced until further
discussion.

- BanKeepAnnotation bans the use of androidx.annotation.Keep.

-BanParcelableUsage prevents any future implmenetations of
Parcelable and directs the user to use VersionedParcelable
instead.

Test: manual testing and ensuring that the error prone errors
trigger on the intended cases.

Bug: 112764060, 115353070, 38123251
Change-Id: I4f02bb270439c47418e0743bfe6b608d966d61a1


GitOrigin-RevId: 17c92cbbf1051eddc01f64ad5c4d87d5d17b8117