Avoid build server tasks in IDE
Recompiling already executed task takes ~ 7s because we sped 6
configuring projects. Configure on demand does not work because
of the way we setup a bunch of check tasks. This CL disables them
when gradle is invoked from IDE since they are not necessary
and then configureOnDemand works fine, bringing down the same
compilation to 2s from 7s.

Bug: n/a
Test: gw :room:i-t:testapp:asAndroidTest --profile -Pandroid.injected.invoked.from.ide=true
Change-Id: I769f3158f5b32f388edceaa17840c439238ddb55


GitOrigin-RevId: 95d6d178758d943177563452a1b69d5e501c07a1