Added support for keeping track of public resource surface
We now detect changes in the public resource surface and treat it
as any other api change, which must abide by the rules of semantic
versioning and require explicit api review when changes are made.

Test: use checkResourceApi task to detect changes in the api
surface and updateResourceApi to update the api. Play around with
the versions and ensure proper behaviour and alerts.

Bug: 77659059
Change-Id: I3d67279b6b5af504a4f0817baee1e59cd2c4c868


GitOrigin-RevId: 86267e31251cdaf875674004b9937ff3da0c3f24