Safe Args in Kotlin - KotlinNavWriter
Finalized the KotlinNavWriter that will generate Kotlin files. The API
for the Kotlin version of the Directions and Args classes is different
than the Java version. The main difference being the lack of builders
and setter. Args and NavDirections classes are immutable data classes
and default values are default parameter values with the idea that
invocations are done with named parameters.

This change also allows users to apply a version of the Safe Args Gradle
Plugin that will generate code in Kotlin. The id of the plugin is
'androidx.navigation.safeargs.kotlin'.

Bug: 110263087
Test: ./gradlew navigation:navigation-safe-args-gradle-plugin:test \
                navigation:navigation-safe-args-generator:test

Change-Id: Ifa0ec7e4bbadd2d274e04dd9ade8a2d619685bd3


GitOrigin-RevId: 07446877340127df5c386e409fbc72a3850a151a