Added checkReleaseReady and checkNoWarnings tasks
checkReleaseReady runs all preliminary tasks to ensure
that a library about to release abides by certain rules.

checkNoWarnings runs the java compiler with the -Werror
flag which generates an error when it encounters a
warning.

This change removes the javac argument "-Werror" from normal
builds and makes it only be present for builds that run
checkReleaseReady.

This helps the prebuilt-drop robot to update the sdk
prebuilt more easily.

Bug:119553356
Test: Manual testing
Change-Id: Id26562d9533d0aaafd3d7cf2e4d9ecab7bccf663


GitOrigin-RevId: ece83c338a46948794d476932dee9e3f43ae27fa