Fixing Gradle deprecated usages.
Got rid of deprecated usages in our gradle plugin or scripts.
There is one remaining usage and that has nothing to do with us,
it's in the kotlin plugin. From what I can see the kotlin folks
are aware of this (https://youtrack.jetbrains.com/issue/KT-26808)
and I'm assuming this will be fixed in time.

Test: ./gradlew bOS
Bug: 122457391
Change-Id: I96125d7c4d1c8380f450625fd5ae253fbeb7f833


GitOrigin-RevId: bb9937c969040d7661b9d30f39ea2c95694c9826