Reenabling Doclava to run against sources too
We're still running Doclava against prebuilts to make public docs, because the output of Dokka for the moment is still different than Doclava.

However, we previously turned off running Doclava against tip-of-tree sources. It turns out that that means that sometimes Doclava would start failing when checking in prebuilt source jars because Doclava is doing different validation than Doclava is.

So we're reenabling running Doclava against tip-of-tree sources too until something changes.

Bug: 72330103
Bug: 122534031
Test: ./gradlew tipOfTreeDocsTask
Change-Id: If7ddc727ffca3e8ff58bf53ec601619e3decf2da


GitOrigin-RevId: a3d28a7904e98d20f1ebd6f97ad22207aa5a9583