Centralize MATERIAL dependency
Remove references to SUPPORT_DESIGN and switch
to the full dependency instead of the previously
used @aar dependency.

Also removes the no longer needed exclude_for_material
exclusion since all libraries are now on AndroidX.

Removed ViewPager2 implementation on AppCompat,
replacing it with an androidTestImplementation since it
is not actually used in the library itself. This fixes
an error when including a ToT implementation dependency
along with the fixed dependency from MATERIAL that
appears when switching to the full dependency.

Test: ./gradlew bOS, ran all affected sample apps
Fixes: 128632612
Change-Id: I44579630cdc3eae58fc23924d8c4d45d2d457f2b


GitOrigin-RevId: 7bbbab7dca4edda160e6b2169d38a25b81698f47