Support Kotlin Coroutines Flow as query return types.
This CL adds QueryResultBinders and Providers Flow. Similar to Rx the
strategy is to generate code that calls into a runtime helper method to
setup Flow given a Callable that actually executes the query.

Channels are not directly supported and instead we show an error
suggesting to change the DAO function return type to Flow and then use
the neighboring functions to create a Channel. The reason is that we
can't reasonably produce a channel without a CoroutineScope and it is
a weird API to ask users to pass a CoroutineScope in their DAO method.

Suspend DAO function that returns Flow is an error.

Bug: 127328278
Bug: 130428884
Test: CoroutinesRoomTest, FlowQueryTest
Change-Id: If0fc6fdb741735f2e4b6b42015730bbf76e8dc18


GitOrigin-RevId: b4c4a0d60152500ec6485624bf74f769ec6d0fca