AndroidOwner sometimes didn't invalidate the correlated node when the model has been updated.
Originally it was implemented as a map with the model used as a key. And given the models are usually mutable data classes they will change their hashCode quite often. It will make the map to lose the nodes affected by this model.
I extracted the logic which solves the issue for FrameManager into a reusable class and use it for both cases now.

Test: ObserverMapTests added, other tests are passing, demo app is working
Change-Id: I58b057902085e4603ea4b474b9a8a91de300fb7c


GitOrigin-RevId: 0114e078b85764c1a176a55cfe04944c1b0271f5