Update library versions to dev version for June 5th Release
This CL only update libraries with versions that are the same as
androidx-master-dev.  For libaries with versions > than
androidx-master-dev, then no version update is needed in
androidx-platform-dev (aka this CL)

CoordinatorLayout will be updated when aosp/972979 merges to
androidx-platform-dev.  It should be identical to androidx-platform-dev
so shouldn't case a build break.

Source of truth: go/androidx-june-tracker
Maven Index Page

Bug: 131579827
Test: ./gradlew checkApi verifyDependencyVersions && gw bOS
Change-Id: Ie5bec5ea0680b5335b470edfa9408b82c2bcc7c0


GitOrigin-RevId: 795ea93a9b032e6ffc36875c1e5676c0c1007902