Added camera-lifecycle module
The camera-lifecycle module will contain all of the lifecycle aware
functionality for controlling a camera. This is so that users of CameraX
that wish to directly control the camera without using lifecycles they
do not have to pay the cost for it.

This initial commit is for creating the module only.

Bug: b/142836519, b/143224674
Test: ./gradlew camera:camera-lifecycle:build
Change-Id: Iae979c52a6fa5ab3f4d10313d0bf80a4ad501b14


GitOrigin-RevId: 0c88a3f0ef9aecda904226d8bbe808f5385500d0