Include all projects regardless of their release statuses when generating zips.
This is to ensure we don't see missing artifacts in our g3 pipeline
immediately following a release.

As mentioned in b/153293171, the reason behind the non inclusion
of released artifacts in the zip is "We did this because gmaven
pushes would fail if you try to push a library with a version
that has already been published. It is trivial to turn it off
if it is no longer needed."

Since releases now happen through Jetpad and Jetpad pre checks
for already released artifacts, there shouldn't be any issues
regarding that concern.

Test: bOS

Bug: 156646472
Bug: 153293171

Change-Id: I897c6a9d33fcaaecbb40e89e01e67039c9abaa5c


GitOrigin-RevId: 15f34f4a55b7cf5ab518674fe65e6f66fa20bf4a