Use jarjar on preferences.proto to solve the issue with diamond dependencies on proto_lite runtime.
datastore-preferences depends on proto_lite which can cause problems if an app depends on a different version of proto_lite. I've created a datastore-preferences-proto module which generates the proto class, then uses jarjar to create a jar with the proto_lite dependency as well as this generated file. I've also put the protolite dependency classes in our name space. I've also created a PreferencesMapCompat class which converts the IPBEs to CorruptionExceptions.

This results in about a 200kb apk size impact on the datastore sample app.

Logic borrowed from Icing:
https://android.googlesource.com/platform/external/icing/+/refs/heads/androidx-master-dev/build.gradle
https://cs.android.com/androidx/platform/frameworks/support/+/androidx-master-dev:appsearch/appsearch/build.gradle

Change-Id: Ic2bd911ea311e06e13d35a865f1d5be4f0f8f317
Test: all current datastore tests pass
Bug: 163129470
Relnote: package proto dependency in datastore-preferences


GitOrigin-RevId: 08a8246a33334ba7b3b48317560cdf51af701cdd