Fixed crash due to synchronous gesture removal.
There was a bug where if a the dispatch to a
PointerInputFilter caused the PointerInputFilter
to be synchrnously removed, HitPathTracker would
crash by expecting the PointerInputFilter to
still be attached.

Also fixed a failing test not related to this
change. Strange...

Relnote: "Fixed crash that could occur when
dispatch to a PointerInputFilter could cause
the PointerInputFilter to be synchrnously removed."

Fixes: 164129464
Test: ./gradle compose:ui:ui:test
Test: ./gradle compose:ui:ui:connectedCheck

Change-Id: I58f253629761b39591e5fdf5b8553d352287d11c


GitOrigin-RevId: ec89ff40be67f9c613706ac8523be8b6ba92dd12