Make placeholder animated and label optional
Before this change label was a required slot API. But often label is not needed, and having it non-optional sends a wrong signal. With this change label is optional similar to other slots inside material text fields.

Placeholder animates on focus/unfocus as per MD specs.

Test: material tests passed + checked the demo
Fixes: 161519460
Bug: 162234081
Relnote: "Label became an optional parameter inside TextField and OutlinedTextField"
Change-Id: I267f6ada96a3371aaa99bdaa4007229ab7efddab


GitOrigin-RevId: 9e43642ea25a9c3f3a63cc7ddd6ef7eafc6348f0