Revamp resistance API in Modifier.swipeable
This introduces a resistance param in Modifier.swipeable that allows
developers to specify not only the resistance factors at min/max as
before but also the resistance basis. The default value of the basis
was changed from the size of the component to the distance between
the min and max anchors, so Modifier.onPositioned is no longer needed.

Test: Ran SwipeableTest and all tests depending on swipeable
Relnote: "The params resistanceFactorAtMin and resistanceFactorAtMax
in Modifier.swipeable were replaced with a single resistance param.
A new defaultResistanceConfig method was added in SwipeableConstants."
Change-Id: I54238306c8356145c8d03507ff20d79cfdab5707


GitOrigin-RevId: 3d1fbc4d9945059ad9d033c9e5390814f950cad4