Improve layout tree consistency checker
Putting the whole logged tree in a error message is not great as on a large trees it cuts. But we can just use println() instead of Logs we used previously.
Also fixed the inconsistency checking rule as it was false-alarming. It is allowed for a child to be in NeedsRelayout state when parent is currently in Measuring

Test: tested locally, the checker was incorrectly failing on in a correct case before this change
Change-Id: I33a803e29abeb9f6688978a541b3aa542d6b73a9


GitOrigin-RevId: 5afc9cd8270d79c66bbdd6968b2a4527ee52c446