Add copy/cut/paste accessibility actions to textfield
This CL shows copy/cut edit options where there is a text selection.
Cut option is shown if there is a text in the clipboard.
Also when the 'Select All' action is trigerred, the floating toolbar will be shown.

This CL does not add the long press.

Test: manually tested with the TalkBack on
Test: copy/paste/cut tests in TextFieldTest
Relnote: "Added Copy/Paste/Cut accessibility actions"
Bug: 154100059
Bug: 153645139
Bug: 167729621
Change-Id: I6db4f570596e65c2e12fbc6f0821961c65671e98


GitOrigin-RevId: b12bf47c6b245d7e3ade043a34ca756499396012