Rename Stack to Box
Relnote: Stack was renamed to Box. The previously existing Box will be deprecated in favor of the new Box in compose.foundation.layout. The behavior of the new Box is to stack children one on top of another when it has multiple children - this is different from the previous Box, which was behaving similar to a Column.
Bug: 167680279
Test: built and ran
Change-Id: I94893bca003d7826c6a5b3c05ac3878d2f6bf953


GitOrigin-RevId: 60a90cc364b6252471973b96fd62abdaddb38d29