Fix SnapshotStateList validateRange helper
The validateRange function should check if an index is within valid list bounds,
starting at 0, and ending at the size of the list (exclusive).
However, the Kotlin '..' operator includes the second parameter as well,
so the range is *inclusive*.
By replacing it with a call to until, the validateRange function works
properly now.

Test: SnapshotStateListTests.validate_subList_get_outOfRange
Fixes: 169471813
Change-Id: I09767bedfad761d33efe09e615efe3eda33c1f00


GitOrigin-RevId: eb129f4202505a4cef2a70e82195a31f5fdeb012