Produce artifact for ui-android-stubs.
This CL changes ui-android-stubs to produce an artifact
so that internal compiler tools will have an easier time
compiling and linking. It also adds a proguard rule to
ensure applications compiling with the stubs don't run
into problems with the missing RendreNode and DisplayListCanvas
classes.

Relnote: "Added stubs for android classes that are on
older platforms to avoid use of reflection when possible."

Bug: 168127161
Test: created and ran app with minify enabled
Change-Id: Ia179051916f28a9924ac9ad733f4b3a2ff6f9844


GitOrigin-RevId: 754a13727a4325a17c773443f8d5ec9227709d56