Rename launchInComposition to LaunchedTask
Compose API guidelines call for Unit-returning @Composable functions to
be named as noun-entities in PascalCase. Rework launchInComposition to
fit this guideline.

Use this opportunity to remove the custom CompositionCoroutineScope in
favor of the standard CoroutineScope; the original purpose of the
nonstandard scope was to carry the frame clock as well, but the clock is
now a CoroutineContext element so this is unnecessary. awaitDispose no
longer fits.

Test: reworked SuspendingEffectsTests.kt

Relnote: "launchInComposition renamed to LaunchedTask to match Compose
API guidelines"

Change-Id: I99a8ef39b1e1abd7b9cae898863a35ed71b62e48


GitOrigin-RevId: ccaf392ee6de89ff533c3c064acc01f6b7df8bc8