Use IR infrastructure for searching symbols
Part of 1366020. Removes usages of symbolTable and uses IrPluginContext.reference# API instead. This change also removes most of the duplicated IR declarations that were created as a result of that and necessity of rebinding symbols. Couple of remaining usages of `generateSymbols` are still required due to weird behavior of `deepCopy` (leaving some typealias references unbound) and bug in compiler fixed in 1.4.20.

Test: ./gradlew :compose:compiler:compiler-hosted:integration-tests:testDebugUnitTest
Change-Id: Ia2223654a55e2c4a94e7d62cf29c27f481daecca


GitOrigin-RevId: 3178e4bb7b876abb448c652251d7d991790cbd55