Fixed the bug where the LazyDSL is crashing on empty list
Fixed a bug where the items and itemsIndexed methods were adding an
IntervalHolder in the LazyListScopeImpl even if that interval was empty.
Thus more than one interval had the same startIndex, which caused the
search method findIndexOfHighestValueLesserThan() to return an invalid
interval.

Relnote: Fix for the items and itemsIndexed methods to allow emptyList
Fixes: 169229639
Test: manually verified the demo app, added tests
Change-Id: I066470185765d918c5f29e96b5ac6dfefe46cb2c


GitOrigin-RevId: d7260d093bf5d59682d860a1e0b7c7c46f847e96