Make the order of place() calls to control the drawing order
Before this change the order in which children of the layout were composed was defining the drawing order. Now it only affects the order of items in the list of measurables received by the Layout. But the drawing order can now be adjusted by the Layout implementation with the order in which it calls place() on placeables. The default layouts like Column or Box are placing children in the linear order so it is a breaking change only for the custom layouts or layouts like SubcomposeLayout where previously the order of subcompose() calls we defining the drawing order. This order is defining the drawing order for children with the same zIndex, so the order still can be overridden using Modifier.zIndex()

Relnote: The order of place() calls in custom Layouts now defines the drawing order for the children
Test: new tests in DrawReorderingTest and SubcomposeLayoutTest
Change-Id: Ibc9f6844b7309f45a8f3dadfdcda0a33b39425e6


GitOrigin-RevId: fc91b887c292b4e6236415429d02b96ca155f80f