Make popup recieve outside clicks.
Because popup was not outside clickable, it was not possible to dismiss
it, if it was not focusable (as clicks outside were not recieved). On
legacy popup this is by default off but since the developer is in charge
of the dismiss callback it makes sense to turn it on and let the popup
user to decide on dismissal. Worth noting that even the popup consumes
clicks outside the system doesn't consider them consumed and propagates
them to the others windows too (if not focusable).

I have also cleaned up tests (and moved some of them to host side to
save us some time in presubmit).

Bug: 164505887
Test: Added dismiss tests
Change-Id: Id5c417d8063a25c94977192ae5277ce720f4e8c3


GitOrigin-RevId: 974771311229743d4ad4d0851030f1c011ed11ec