Fix bug in shadow alpha.
setHasOverlappingRendering(true) was incorrectly removed in a
previous CL, so this has been added. I validated the
opacityAppliedForTheShadow() manually -- verifying that
the luminance differs when opacity is 1 vs 0.5. It appears
that the actual luminance differs between versions and
devices, so exact numbers can't be used. I used several
devices to check ranges and opened it up so that it will
pass on a wider range of devices.

Fixes: 169835769
Test: DrawShadowTest.opacityAppliedForTheShadow()
Change-Id: I4bd377ea455d7a55312538c2c9111fd2a1fca427


GitOrigin-RevId: f374f4344f02970a5d5e42c80ff0f53d70cc4c08