Add SideEffect API to compose-runtime
SideEffect is intended to eventually replace the onCommit effect.

A SideEffect of composition will run its provided lambda when the
composition successfully commits. It is not a
CompositionLifecycleObserver and is not stored in the composition's slot
table. It does not run conditionally, and is intended for applying
changes to external objects managed by the composition that are not
backed by snapshot state that would benefit from the snapshot commit
mechanism for synchronization. It does not provide a cleanup mechanism;
use cases requiring cleanup and lifecycle awareness should remember
their own CompositionLifecycleObserver object.

Relnote: "Added SideEffect composable for applying side effects of
composition to objects managed by the composition. SideEffect is
intended to replace the onCommit composable."

Test: SideEffectTests.kt
Change-Id: Ia77c2060e10aa8011052e18ece011297fc28831c


GitOrigin-RevId: 0abdd94627e6900352313f8e72f956b986159543