Add DisposableEffect and rememberUpdatedState APIs
DisposableEffect is the counterpart to SideEffect and fills the role
that onCommit-with-params plays. In contrast to onCommit,
DisposableEffect's onDispose clause is required as the final statement
in the provided block by way of returning an inlined interface
implementation for a disposer. This is a soft requirement; no
implementation restrictions are placed on constructing objects
implementing the required DisposableEffectDisposable interface returned
by onDispose to allow library implementations to plumb implementations
through as needed.

DisposableEffect does not permit usage with no subject parameters,
directing callers to provide one by way of a deprecated overload
shadowing the varargs overload that would otherwise permit this usage.
Callers are instead directed to SideEffect for these use cases,
discouraging onActive-like uses.

rememberUpdatedState is useful for working with DisposableEffects and
LaunchedTasks that configure ongoing processes or coroutines that are
prohibitive to tear down and restart if input data changes.
rememberUpdatedState offers a concise way to update data in
recomposition and have those changes visible to long-lived objects that
closed over the resulting state holder.

Test: SideEffectTests.kt
Relnote: """Added DisposableEffect counterpart API to SideEffect, filling
the role of onCommit-with-params but with a required onDispose.

Added rememberUpdatedState API to publish data from recomposition to
ongoing or long-lived processes such as DisposableEffects or
LaunchedTasks."""

Change-Id: Id50b9b85d2303166e5abe10aea47d6934d2c8597


GitOrigin-RevId: 2435c3f67c24fefcc182c394d7b99c045d9664c8