Recomposer extends CompositionReference
Unify the relationship of compositions to their parents by making
Recomposer into a valid composition parent on its own. This removes the
need to read/specify a Recomposer in many places and paves the way for
simpler composition bootstrapping.

Relnote: "Recomposer is now a CompositionReference and a valid
composition parent. Explicit Recomposer is now required in fewer places."

Test: existing tests

Change-Id: I4036ff66dff4759bd40127a9af0ca59cbaa37041


GitOrigin-RevId: 30a76a1c821c6fe7c2dee6f4a086445e029b2218