Made ConsumedData a non-data class.
ConsumedData is mutable and is contained within PointerInputChange which is a data class. By keeping ConsumedData a data class, it meant that the equals and hashcode of PointerInputChange would not be stable which is error prone. Now that ConsumedData is not a data class (and does not implement equals or hashcode), PointerInputChange's equals and hashcode are stable despite any mutations done to ConsumedData.

Relnote: "ConsumedData is no longer a data class. See https://android-review.googlesource.com/c/platform/frameworks/support/+/1455219 for details".

Fixes: 170666036
Test: ./gradlew compose:ui:ui:test
Test: ./gradlew compose:ui:ui:connectedCheck

Change-Id: I1737f9928e7731402e896deedf7d8a97a598e725


GitOrigin-RevId: 8172320488330625233e63e510ccb3dea0174b9a