Stop setting zIndex in Surface
We decided to decouple shadow size and zIndex even in material components. Previously Surface was setting zIndex to the same value as the shadow elevation. This behaviour was often confusing and non desireable. It will be more clear if just drawing reordering would be done in the opt-in manner when needed. Users still can pass Modifier.zIndex() when it is needed.

Fixes: 170623932
Relnote: Material components do not set the elevation as zIndex anymore. Which means that within the same parent the child with larger shadow size will not be drawn on top of the child with the smaller one automatically. If you still need such behavior please set Modifier.zIndex() manually where needed
Test: material module tests
Change-Id: I704173218a6e049c215b33d5a71234e0eaba1c44


GitOrigin-RevId: 1b87aeffbfe320e99cfa569ea23971a3b1698b5e