Composition Applier begin/end changes callbacks
Add callbacks to Applier signaling the beginning/end of a batch of
changes to allow for associated batch resource management. Document an
unexpected language behavior that can impact lambda-heavy code.

Relnote: "Applier now has onBeginChanges/onEndChanges callbacks that are
invoked when a Composer begins/is finished applying changes to the tree.
These may be used for batching resource management if needed."

Test: CompositionTests#testApplierBeginEndCallbacks

Fixes: 171578075
Change-Id: Icf4765f3fd6102d26177aac6f5f259f9b9c0c0de


GitOrigin-RevId: 437eb8a97c4fba3b6a3282031c6efb25b870acbe