Adds lintChecks dependency on ui:ui-lint to ui:ui
lintPublish doesn't add the lint check as a check for itself by default, so this means that ui:ui-lint will now run on ui:ui.

Updates ModifierDeclarationDetector to ignore variables (such as inside a class).

Also uses InheritanceUtil in ModifierDeclarationDetector as this is simpler and may avoid infinite recursion in cases where the inheritance chain has a cycle.

Bug: b/172469237
Test: lintDebug
Change-Id: Ia2b88c4b55676194e388cf25fdf0f6a8ef36aac9


GitOrigin-RevId: 192318f2c1de4533d5b21141a78cd57d588452d5