Add setup vs measured block to macrobenchmark
Fixes: 173155412
Test: AppStartupHelperTest, MacroBenchmarkTest, JankMetricValidation

Adds a 'setupBlock' for initialization that shouldn't be measured. For
example, this can be used to navigate to a jank test's scrollable
content.

To avoid measuring this new section, we also transform metric results
and traces to be per-iteration.

As part of writing tests for altered AppStartupHelper behavior, fixed
multiple issues with error clarity:

 * Startup capture failure is escalated to an exception

 * Selinux enforcement detected is fixed, and failure to set
   permissive is escalated to an exception

setupBlock API may still be improved - would be nice to pass values to
measured block, allowing you to e.g. locate RecyclerView in setup,
setup a gesture, and only actually scroll it in the measured block.

Change-Id: Ieaaa12073d9ed1cca5b1c34f9b9141fec14bcb78


GitOrigin-RevId: 0bd7a1cece787fd8b95c5ae77771047853147b79