Fix accessibility for Checkbox/RadioButton/Switch
This adds a API that clears all descendant semantics from the merged
tree, called "clearAndSetSemantics". Although the main point is to clear
semantics (and it can be used that way with an empty lambda), I added
the "set" in order to use the power of suggestion on developers not to
simply drop semantics on the floor but always try to provide an
equivalent replacement.

The immediate use case, which I solved in the samples, is to support a
clickable list item with a checkbox/switch inside it, without creating a
redundant tiny screen-readerfocusable box for the checkbox.  I tried
several of the proposals in Anastasia's design doc and this approach
proved to produce the cleanest code.

The wider use case is to support all manner of app-specific custom
screen reader behavior, in the cases where the desired behavior
conflicts rather than purely adds to Compose foundation's default
semantics.

Test: SemanticsTests.clearAndSetSemantics
Relnote: "Added Modifier.clearAndSetSemantics to clear descendants'
semantics and set new ones."

Change-Id: I277ca1b76fda44a34a2c01844b832244cb42ff7e


GitOrigin-RevId: 0367a37bd9501344e210dc740512eb93c4170456