Move Composable property annotation to getter
Fixes: 171981332

Relnote:"""
Compose supports property getters that can make composable invocations. Support for this is not going away, but the syntax for declaring a property getter as being @Composable is changing.

The now-deprecated syntax for doing this was by annotating the property itself:

    @Composable val someProperty: Int get() = ...

The now-correct syntax for doing this is by annotating the getter of the property:

   val someProperty: Int @Composable get() = ...

Both syntaxes will work for some time, but the former deprecated syntax will eventually become a compile error.
"""

Change-Id: Id91977f2583b81d3e4e51bbf120cfaf943be25d5


GitOrigin-RevId: 8de8feff85ad20f6006a50f067d5bd16904ecfa9