provide SemanticsActions when a SemanticsNode is marked as disabled
In android, we have a clickable property. when disabled, we don't
add the click accessibility action (we still have the onclick
listener though for non accessibility), but we still set the
clickable to true.

In talkback, we only notify disabled when the node is actionable.
In android, since we have clickable set the node is actionable.
but for compose, clickable is not set, so talkback doesn't announce
it is disabled because the node is not considered actionable.

This cl change it to always add semantics actions, even when the node is
disabled and convert to corresponding android accessibility properties.
For more details on different solutions considered, please see the bug.

Relnote: Add Toggle to foundation Strings.kt

Fix: b/172366489

Test: tested with demo app and talkback. Added a disabled toggle button
in the demo app to make sure disabled nodes can't be toggled. Added an
integration test to make sure disabled nodes can't be toggled.
Unit tests added.

Change-Id: I4a5b74e3ed9bc3b1acd09af221331ef6ab51b9fe


GitOrigin-RevId: 1e6d8aa505d2efde5f61ff3c45325ab533840d32