Add support for deferred brush creation
Relnote: "Refactored ShaderBrush to
lazily create a shader instance when
sizing information of the drawing
environment is available.
This is useful to define gradients that
occupy the full drawing bounds of a composable
at composition time, without having to
implement custom DrawModifier implementations.

Deprecated gradient function constructor APIs
in favor of factory methods on a Gradient object."

Bug: 173066799
Test: Added tests to DrawScopeTest
Change-Id: I511fc09bdeb4b41127de4a6b1e616688b10295f5


GitOrigin-RevId: e9f976c645e2c449df86e2a22e3e43da2d3595b9