Leverage TestCoroutineDispatcher in testing
This CL adds a new ComposeIdlineResource that relies on the
AndroidComposeTestRule to inject TestCoroutineDispatcher. With this we
can advance frames until we are idle. This also relies on a trick that
we inject our own animation clock that are subsribed to the
MonoticFrameClock without this the animation would still run on
Choreographer and this solution would not work.

There is still a legacy way preserved by using createTestRuleLegacy.
Hopefully that is not something that will make it to release as we
migrate all the remaining tests before that.

Test: Pass
Bug: 168695905
Relnote: "Leverage TestCoroutineDispatcher in testing"

Change-Id: I532b68e37ea6f72964fdcc267e397d285cffd9ad


GitOrigin-RevId: e9a867012d1f181d8fdf8c62d95b6e462001da6f