Prevent OOM in measureCompose runtime benchmarks
This switches measureCompose to be suspending and manual-time-control in
order to call advanceUntilIdle() after composition.dispose().  Giving
the dispatcher an opportunity to run is required to fully clean up
memory after cancellation of the LaunchedEffect I added in aosp/1435685.
This is a similar change to the one I did in aosp/1519221 for a
different set of benchmarks.

Based on my local Wembley testing, this will cause an apparent
regression in all the affected benchmarks (a large regression in the
case of the simplest benchmark cases like OneRect), but not to introduce
additional noise.  So these benchmarks should remain equally useful if
we accept a one-time discontinuity and compare to the higher baseline.

Bug: 175626893
Test: Existing tests
Change-Id: I6f2329bcc281d6ab8c7ece5e10eec43fe0200aec


GitOrigin-RevId: a43b92f160d71b965c67b29ab6b2a6dd9e0535c3