Add decorationBox to the CoreTextField
Scrolling logic moved to the CoreTextField with gestures part applied to the decoration box and layout part of scroll applied to the internal text field implementation.

Disabled text field and readOnly text field implementation moved to the CoreTextField.

Fixes: 171484101
Fixes: 171479319
Test: added a few new tests
Test: foundation and material tests passed
Relnote: "BasicTextField received a new parameter called 'decorationBox'. It allows to add the decorations like icons, placeholder, label and similar to the text field and increase the hit target area of it."
Change-Id: I169960830df82b0406ac4b6868ea544c9f848403


GitOrigin-RevId: e0c52d3e1434fc26069034291eebee6d4b61eda8