Create nodes and set value only during applyChanges
Prior to this change composition created nodes and set
properties on these nodes during composition but the
node tree was assembled or structurally changed during
applyChanges().

After this change nodes are created, initializerd,
assembled, and updated only in `applyChanges()`.

This is an initial step for more flexible threading as
nodes are never read and only written to during
applyChanges().

As before, nodes are not read during composition, and
after this change they are no longer written to during
composition (e.g. when any `@Comoposable` is executing).
All changes are deferred until the apply changes phase
is executing.

BREAKING CHANGE: Nodes are no longer available during
composition. This required some changes to tests that
assumed the node had already been created. These tests
were fixed by wrapping the code in `SideEffect` which
schedules the code to be applied along with the other
changes to the nodes.

Similar changes might be required for applications
that make similar assumptions.

Test: existing tests pass
Fixes: b/160612407
Relnote: """Restructuring of the internal compiler
API allows batching changes to the nodes generated
as a result of  composition into the "apply changes"
phase of composition, after all @Composable functions
have completed.

BREAKING CHANGE: This is a behavioral breaking
change that might affect application code as
nodes are no longer available from internal and
experimental APIs until after changes have been
applied. This can usually be worked around by
surrounding code with such dependencies in a
`SideEffect` composable to defer execution of the
code until after the nodes have been created and
initialized."""

Change-Id: I018dab05a0486e8db663aea39a7546aa73142c11


GitOrigin-RevId: a5166d60e322403beb28f5baea18a4160677075f