Fix issue where clicking on a non-clickable area didn't request view focus.set root focus correctly.
When a user clicked on non-clickable space in the app, we cleared any existing focus.
However if nothing was focused, we just set the root focus modifier to Active without
requesting focus for the owner view.

This issue was manifested by text fields that would appear focused but would not show
the keyboard (because the owner view was not in focus).

This CL includes a fix and adds a unit test.

Bug: 170188271
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.OwnerFocusTest
Change-Id: Icd9e737b2a977dd1f8d6219bba8967a8b1b34fca


GitOrigin-RevId: 262f1d6ee5e93f7a0a8e6224e84596c43d338c29