Changes Material stateful parameter interfaces to have @Composable functions that return State<T>
This is consistent with recent changes to animation (animateAsState) and also ties these functions more strongly to the composition, since the animations use the coroutine scope tied to composition.

Adds Animatable.asState() to make it easier to convert an Animatable to a State.
Also changes animateElevation to be a suspend extension on Animatable.
Also migrates RadioButon to no longer use animate(), and adds support for benchmarking components with suspending animations.

Bug: b/146345816
Fixes: b/176192329
Test: updateApi
Relnote: "Changes Material stateful parameter interfaces to have @Composable functions that return State<T>. Adds Animatable.asState() to make it easier to convert an Animatable to a State. Also changes animateElevation to be a suspend extension on Animatable."
Change-Id: If613cc7c751a11b77a03f8066b233b7e55cb67e0


GitOrigin-RevId: ccd5081949dce63408b532cb3b3f90474933f684