prune accessibility tree to remove nodes covered by sibilings
If a node is completely covered by its sibilings
(we don't prune parents because a touch
filter in children can still pass touches to parents, like the parent
can be scrollable), this node is removed. The detailed reasoning can be
found in aosp/1532973 (dependent cl).

I'm not sure whether it is a bit risky. For all-text nodes, if
we have a transparent sibiling on top of it handling touches (thus the
sibiling will show up as a semantics node), the covered text (still
visible) will not be accessibile. For example, a book annotation app
which has a transparent layer on top of text and click on a location will
create a note on this transparent layer. But such an app is not
accessibility in nature (at least to talkback users and switch access
users) and we need to ask developers not to report this transparent
layer. But I'm not sure whether there are other common usage cases where
prune the tree this way will result in loss of information. If we find
problems later, we can revert this cl.

Note that pane api implementation in
AndroidComposeViewAccessibilityDelegate.kt depends on this cl, so
if this cl is reverted, the pane api implementation needs to revised.
Here is a reference cl(does not depend on tree prune)
if this needs to be done: aosp/1482398.

Test: tested with drawer, bottom sheet, and backdrop (with proper
semantics to report expand/collapse), and the covered
items can't be accessed (both linear navigation and touch explore) in
talkback. Unit test added in AndroidAcessibilityTest.kt.

Change-Id: I7a48d5410470365c29e3bd3e7ba6dff033607b05


GitOrigin-RevId: fce5a3186698e6b22248cfa2f0521ae1a15d677a