Add support for converting function reference instances
Function references can mostly be handled by the existing support
for lambda instances. On the Studio side the PSI for locating a
function reference is different from a lambda expression.

Instead of just passing the instance (of a lambda or function ref)
we now are passing an array which also will contain the name
of the function being called. This will be used on the Studio side
to find the function reference in PSI.

Bug: 177344904
Test: Added test for function references
Relnote: Added a function reference parameter type
Change-Id: I5e1bdbe66deb2528a7e152113818a2d7ff5303e6


GitOrigin-RevId: 3df27d53f2521002237fa84491228f537494222f