Set ComposeView's lifecycle owner before attached to window
After the recent migration from Recomposer.current, ComposeView is using a
window-scoped recomposer, which requires a ViewTreeLifecycleOwner to be set
when attaching the view to window, as this recomposer will ensure that
composition is created on attach.

In order to prevent the crash that will ultimately happen when checking if
the Compose Preview's ViewTreeLifecycleOwner is not null, we simply set the
ComposeView's root lifecycle owner *before* onAttachedToWindow().

Test: ComposeViewAdapterTest#previewInClass
Fixes: 178354528
Change-Id: If44d3a1c8bdeca2bddbda2f6c43b1855c6a33f09


GitOrigin-RevId: 636d61822bc3b66cbd0d8a24c921381914c26f7d