Deprecate resource loaders in favor of painterResource
As agreed in API review, this patch deprecates all resource loading APIs
except for fontResource, painterResource, and the primitive resources
(string, color, etc).  The deprecated image/vector APIs APIs were
redundant with painterResource, and the bulk of the patch is that I
changed samples and unit tests to use painterResource.

The rationale for the deprecations is:
1) loadFontResource is thread-unsafe if API < 27. The small potential
performance benefit is not worth that potential timebomb.
2) imageResource and vectorResource are redundant with painterResource.
painterResource is a safer API we want to encourage developers to use
because it will tolerate resource IDs being converted between bitmap and
vector without crashing.
3) There is a long-term need for an asynchronous loading API, but it is
not urgent (partly because resources are usually small, and partly
because this feature can be implemented by outside devs and we only
offer standard resource loader API as a convenience).  API council is
not satisfied with the existing DeferredResource approach to
asynchronous loading, so let's delete it before we're stuck with it.

Bug: 174573385
Test: Existing tests
Relnote: "loadFontResource is deprecated.  Use fontResource instead.
imageResource, loadImageResource, vectorResource, and loadVectorResource
are deprecated.  Use painterResource instead."

Change-Id: I6b8096508b2280ca49c70a432c5497f386dc570e


GitOrigin-RevId: cad5eb8ea7ee46610d9f555bd50ab79745e1d6c5