Remove AnimationClockObservable and AnimatedFloat
RelNote: "AnimationClockObservable and subclasses have been
removed. AnimatedFloat has been removed."

Note: this CL is based on aosp/1576426 and aosp/1570862, but
the parent pointer isn't pointing to either of those. This is
to prevent the rebasing of this CL from accidentally uploading
patches to the two CLs above, creating unnecessary churns.

Bug: 177457083
Bug: 178448485

Test: removed
Change-Id: Icde5248072620b741bdf4d8cf59291e7b2994e6a


GitOrigin-RevId: 0a55e899914b1398bed87b33303b083b2efcd418