Make compositionLocalOf default value function required
Relnote: """The `defaultFactory` for `compositionLocalOf` and
`staticCompositionLocalOf` is now required instead of
optional.

This changes removes a potential type error for non-nullable
types where no default factory was provided. Previously this
would provide a null reference for a non-nullable type.

For nullable types consider supplying `{ null }` as the default
factory.

We do not recommend using locals with non-nullable types unless
a sensible default can be provided. If no sensible default exists,
the `defaultFactory` lambda should throw an exception. However
throwing an exception means that consumers of the local will have
an implicit dependency on it being provided that is not enforced
by the type system."""

Test: ./gradlew bOS
Fixes: b/172939815
Change-Id: Ifbd2a8abd2f1fc4578da4b782b570ed9de088beb


GitOrigin-RevId: ba897037733e30bb8fda45c7bd85a6c898d501c7