Make TextFieldValue constructor public
In order to be able to test TextFieldValue from
foundation we need to make the constructor with
composition public. Previously we hid it because
developers would create a new TextFieldValue
with composition set to null however it wasn’t
a big improvement and with the current internal
constraints it isn’t worth it.

RelNote: N/A
Test: Treehugger
Bug: 179710063
Change-Id: I0bda53cd3f2e2cbf9ae5d8b6bc4fdb7c29714a7f


GitOrigin-RevId: c5ccee45bd39d17916944d5166280a73968cbf93