Make EditProcessor public
EditProcessor is in ui-text and used by foundation.
It is currently marked as @InternalTextApi. However
since it creates a inter-group dependency we have to
change it.

In this CL, EditProcessor is made public so that it
can be used from foundation.

EditProcessor uses a set of functions from EditingBuffer.
It is possible to move EditProcessor to foundation, and
make EditingBuffer public.

Unfortunately neither of those classes had feedback
from either developers or API reviews. EditingBuffer
would result in more unverified API’s therefore this
CL choses to make EditProcessor public.

RelNote: N/A
Test: Treehugger
Test: ANDROIDX_PROJECTS=COMPOSE ./gradlew jvmTest \
  desktopTest -Pandroidx.compose.multiplatformEnabled=true \
  --no-daemon

Bug: 179710063
Change-Id: I252d5b6509298f3aed846dd812507b444e8f72a3


GitOrigin-RevId: 825fc2f7f6ecc13164b859d38d2468a3278f1532