Improve selection control APIs to support "passive embedding"
This makes the Switch/Checkbox/RadioButton lambdas nullable.  If null,
then the entire "toggleable" modifier is removed, disabling click input
handling, Material ripple and accessibility semantics.

These are intended for use cases (as in the samples) where the control
is embedded within a larger row which handles the click action for it.
Note that this is distinct from "enabled", which is for use cases where
the UI is grayed out to indicate it cannot currently be interacted with
in any way.

Bug: 171819073
Test: "New 'whenNullLambda' tests in RadioButtonTest,
SwitchTest, and CheckboxUiTest"
Relnote: "Switch, Checkbox and RadioButton action lambdas are now
nullable.  Checkbox-in-clickable-row samples updated to use this
feature."

Change-Id: If601b88cf4622111bca5f4927cbb86c7d300ebbf


GitOrigin-RevId: a234d18045d35a211dfceea56b20fece51dc34cb