Remove SoftwareKeyboardController
This change removes the callback APIs that were previously on
BasicTextField, and all Material TextField implementations. Calls to
control the keyboard are not possible via any high level APIs after this
commit.

A new API to control the keyboard will be added in another CL to follow.

Bug: b/168778053
Relnote: "Removed SoftwareKeyboardController callback from all text
fields to be replaced by a new API shortly."
Test: Automated, manual, and demos

Change-Id: Iae869e91c48300f4ab926dac2578d2d759f5fd89


GitOrigin-RevId: d46aa1d604c3c3177e084652bdc234401dc60e74