Changes to support constraint visualization in Studio
Adding a way to retrieve Constraints information from the
ComposeViewAdapter. Which can be used to visualize them in the Android
Studio Preview.

While this is meant to be used primarily for ConstraintLayout
visualization, this implementation really only looks for the desired
method signature and could be used to support other libraries.

Test: ComposeViewAdapterTest
Change-Id: I9ce28575cedd9df88704159a1455605362d27bce


GitOrigin-RevId: f627f56981dbd9912f8ee4fe12076bd0739c54b5