Desktop: fix zero size on the first frame drawing
After we implemented DirectX we discovered that our first frame renders with zero size.
So after the first frame we will have transparent window.
We noticed this because the second frame was too long.

It happens because our listener "componentResized" fires only after initialization of ComposeLayer (when it adds to the UI root).
And on initialization we draw the first frame.
We need to implement "setBounds" instead, because "setBounds" is called immediately before rendering.

Test: manual ./gradlew -Pandroidx.compose.multiplatformEnabled=true :compose:desktop:desktop:desktop-samples:run1,
Test: manual ./gradlew -Pandroidx.compose.multiplatformEnabled=true :compose:desktop:desktop:desktop-samples:run4,
Test: Should be successfully started and should resize content on window resize
Change-Id: Ifd3aa67a0a301212aac288c6ffcf1d402fefd233


GitOrigin-RevId: 74399ed2fe864236498d677d4cd69707d3f853fe