Incremental processing for KSP
This CL adds incremental compilation for the KSP backend.

We do not respect the `room.incremental` flag since it was added to
initially enable incremental java annotation processing and KSP is a new
backend hence we should always be incremental. (developer can still turn
it off for all of KSP).

The way we collect these files is a bit hacky because it requires us to
wrap a KSFile into a javax.model.Element because the only API available
in JavaPoet receives javax.model.Element. An alternative implementation
could abstract JavaPoet but it would be a more complicated change.

There is a KI in KSP: https://github.com/google/ksp/issues/332
* KSP does not delete previous outputs if originating files are deleted.
That test is disabled for now until there is a new KSP release with the
fix. This does not easily reproduce in a sample app but does reproduce
in our testing setup.

Bug: 176453350
Test: RoomIncrementalAnnotationProcessingTest
RelNote: Added incremental compilation support for KSP.

Change-Id: I031c1f94890ebe2a382c26f7f0745edb790a5a7b


GitOrigin-RevId: fc9df1181ed75a213c28b917d96adc4076163aa7