Refactor LazyListScrollingBenchmark
We previously didn't do real scrolling in the benchmark as we had no api for it. Now instead of resizing first item we do real scroll bechmarking a real world use case. We also previously didn't disable tracking for `assertNoPendingChanges()` and we need to do so as it takes some time when we have state mutations and it is not really a part of Lazy layout logic. Linked regression was reported because of it: previously we didn't have state value change in the logic because scroll offset didn't really change and after exposing LazyListLayoutInfo we now had the state change. It is incorrect, as we will always have state changes when scroll happens.

Test: run benchmarks on wembley. the numbers will change as we changed the exact logic being done inside the benchmark block
Fixes: 175592218
Change-Id: I47d120e1140b94b17ab5608d7d7b0f5f43dc06a8


GitOrigin-RevId: 714bb7edc4ced54e02446e614318ff8db794a7b2