Get size of semantics node differently
If the node is merging, than outer most semantics modifier that has mergeDescendants == true will be used to get the bounds for semantics node. Or otherwise the outermost semantics in the chain will be used.

Along the way fixed a few bugs introduced after globalBounds have been deprecated and replaced by boundsInWindow.
In the acessibility scrolling optimised for not using the boundsInWindow to get the width/height but kept the logic where innerLayoutNodeWrapper is used to exclude the padding.

Test: ./gradlew compose:ui:ui:cC
Test: ./gradlew compose:foundation:foundation:cC
Test: ./gradlew compose:ui:ui-test:cC
Test: ./gradlew compose:material:material:cC
Test: demo with Talkback, Select to Speak on
Test: new tests in AndroidAccessibilityTest
Fixes: 180585803
Change-Id: I3175dc4b872cb6a94bb6de0f71ac2524d356ef4b


GitOrigin-RevId: 5af3d148a5149b5d3883929694e1ceff1eb411f7