Deprecate DefaultMonotonicFrameClock
Providing a useful DefaultMonotonicFrameClock for each supported
platform that behaves in an acceptable way has proven to be prohibitive.
With kotlinx.coroutines now providing a CoroutineContext.job extension
that throws if a required CoroutineContext.Element is not present as
precedent, add CoroutineContext.monotonicFrameClock that behaves in the
same way.

Relnote: "DefaultMonotonicFrameClock is deprecated. Calling
withFrameNanos or Recomposer.runRecomposeAndApplyChanges with no
MonotonicFrameClock will now throw IllegalStateException."

Test: MonotonicFrameClockTest.kt
Change-Id: I4eb0d7a8ebae7497735d25bc35e9f94c66ce2232


GitOrigin-RevId: 662bf10e07f0f880d71df778e6d9ca6a829c9718