Ensure AndroidViewBinding uses the right LayoutInflater
Fragments rely on nesting fragments when it comes to
saving and restoring their state properly. When using
LayoutInflater.from(context), any fragments inflated via
FragmentContainerView are always attached to the
Activity's FragmentManager.

This leads to issues when your Compose hierarchy is
inflated as part of a Fragment. In these cases, we
specifically want any AndroidViewBinding inflated
fragments to be added as child fragments. We can
accomplish this by checking whether the LocalView
is owned by a fragment and, if so, use that fragment's
LayoutInflater (this is the magic that causes inflation
to inflate fragments into the right FragmentManager).

Relnote: "`AndroidViewBinding` now correctly nests
fragments inflated via `FragmentContainerView` when
your `ComposeView` is within a `Fragment`, fixing
issues with saving and restoring the state of those
fragments."
Test: new testRecreateChildFragment test
BUG: 179915946

Change-Id: I70eb0831fb6b756e2968e713e181193969d49b9a


GitOrigin-RevId: 0934428c2e01d225bbf29485ec539ae60ea88c98