Add error API to mark node that contains invalid input
In accessibility services there're two APIs - setContentInvalid(Boolean) and setError(CharSequence).
And yet setting the error text requires the former to be set to true. At the same time, just setting contentInvalid to true will not result in anything.
Therefore having two APIs is not very beneficial and may only cause confusion. As such, in this CL I introduce a single API that takes an error description.
Name was not a clear choice, but decided to go with `error()` (as oppose to invalid or invalidContent) because
* this is what TextView uses, so developers should be familiar
* our material text fields have a 'isError' parameter for styling purpose. Therefore, it will be easier for developers to find the semantics if it has a very close name to something that is already treated as error.

In material text field we supply error semantics with default message in case developers forget to provide this semantics.

Relnote: "Added accessibility API `error` that allows to mark a node that contains invalid input"
Fixes: 180584804
Fixes: 182142737
Test: new tests in AndroidComposeViewAccessibilityDelegateCompatTest
Test: manually with Talkback on and alternative locale
Change-Id: I129977e3d3b5f03435de41de409fde9029c325b9


GitOrigin-RevId: 72e2c2a4cddeaf9a80e664fd6fb8a12b9d962c48