Update KSP to alpha 06
This version has 2 important changes.

In Alpha 05, * projections in kotlin source files were not resolved to
any type hence we were able to detect them as is. Appearantly, it is not
how it worked for .class files but we didn't have tests covering it.
In Alpha 06, KSP delegates to the compiler in more places which means a
type declared as:
MyType<*> actually gets resolved as TypeReference(variance:START,
type=Any?). Unfortunately, it is exactly the same way it is resolved
when declared as MyType<T> when T is not specified. Ideally, resolving
these to the upper bound would be more correct but that is not what
happens with JavaAP `TypeElement.asType()`. As a result, I just removed
the test where we expect `MyType<*>` to match KAPT and also expanded all
KSTypeExtTests's to cover compiled files. This change makes it more
consistent within KSP even though it slightly diverges from Java AP.

Another important change is that KSP fixed how annotation use site
targets are handled such that annotations with use site targets show up in
the target element's annotations (that is what happened for .class files
already). If target element is implicit, KSP still generates it as a
synthetic. Note that, even though this reduces the complexity of
KspAnnotated (no combine logic), we still need to check for use site
targets and the current bug where we don't find default use site target
is still present. (has a TODO for it already).

Bug: 160322705
Test: existing tests
Change-Id: Ia35dd2def79fb5aed983f5153d4f4e051aefcebb


GitOrigin-RevId: 3f9bc79b1260e43bc91fac07a7fb978e7c2b9f21