No-reflection Layer implementation
There are some devices in the wild that don't support
updateDisplayListIfDirty() reflection, so we must
support a non-reflection implementation of layers
in compose. This adds a layer that doesn't use
reflection. The "updateDisplayList()" is replaced
with calling "super.dispatchDraw()", but clipping
out all drawing so that none of it reaches the
screen.

:compose:ui:ui tests have been changed to force use
of this implementation for P devices. L devices
will continue to test the previous implementation
of ViewLayer.

Fixes: 171985186

Test: ran all :compose:ui:ui connectedCheck
Test: visual check of demo app, including animations

Change-Id: If0ea7a8f9a41d4e7b5aafcb111bf3ab3a4a9454b


GitOrigin-RevId: 4c8dbb5b7efefbebb098d1c0d7a4ad0787127b1e