Fix problem with pointer input having a bad position.
Fixes: 169874631

In some circumstances, getLocationOnScreen() can return an invalid
result during pointer input dispatching. This CL uses the motion
event's raw position and position in the view to determine the
window position. It also calculates the view matrix once and caches
it so that it won't have to be calculated multiple times during
pointer input dispatch.

Test: ran breaking application.
Change-Id: Ic99fc8b4628b92d79f76da6e46363cebbd95ee4f


GitOrigin-RevId: 03b281a865ebeba81087864dbccc4476ec071dc4