Add Collection(Item)Info semantics API
Any lazy list by default will report "-1" as the number of elements because we can't calculate correctly the number of elements that should be reported for a11y.

When `Modifier.selectableGroup()` is used with lazy lists, we won't report anymore only visible elements on the screen.

Test: new tests, manually with Talkback on
Bug: 180479017
Relnote: "Added CollectionInfo and CollectionItemInfo accessibility APIs that allows to mark collection and its items for accessibility services"
Change-Id: Id54ef37379e14e41ac52782b40e29de54f95eed0


GitOrigin-RevId: 8dddaeb8d6af1e2c6ae4518376132d8c00d3cf97