Fix memory leaks with state objects
This fixes two memory leaks reproduced by the same test
case (b/182822837).

The first prevented state records from correctly being reused
when a state object is modified due to typos in calculating the
lowest open snapshot.

The second prevented objservable objects from being correctly
unobserved when the recompose scope was recomposed but was not
invalidated and that scope stopped observing an object.

Fixes: b/182822837
Test: ./gradlew :compose:r:r:tDUT
Test: ./graldew :compose:r:r:i-t:cC
Change-Id: If96bde0c7b9172918b6d641a656c1fe15132ba4b


GitOrigin-RevId: d5f990aa941647aae11092b47aa550caf31c9bc0