Add scroll to matcher action for LazyLists in tests
Adds an action to SemanticsNodeInteraction to scroll to content in a
LazyList identified by a SemanticsMatcher. Adds a filter as well to find
nodes that support this scroll action.

Fix: 161584524
Fix: 182259407
Fix: 182260995
Test: Added ScrollToNodeTest
Relnote: "Added `performScrollToNode(matcher: SemanticsMatcher)` that
scrolls a scrollable container to the content that is matched by the
given matcher."

Change-Id: Ic1cb855e351c7bb683962d618d68782628b70f62


GitOrigin-RevId: 59dca3bb1906b988bcc55a2bc44f93063e87a48f