Changing the way we determine presubmit builds.
This uses a much better identifier than the previous heuristic.
The key functional difference will be that all local builds will
create configs as if the build were a postsubmit build. You would
need to set BUILD_NUMBER environment variable now, in order to mock
presubmit configs. Since these configs are for test harness ingestion,
and it is not expected that engineers need to debug any difference
locally, this seems acceptable.

Test: Manual comparisson of treehugger build output.
Bug: 182293323
Change-Id: I7fe0b12cf213a44052196e4eba3fea47e93d54aa


GitOrigin-RevId: 0785997cb3edb782082640b60fdadd44b93d5324