Changing the way we determine presubmit builds.
This uses a much better identifier than the previous heuristic.
The key functional difference will be that all local builds will
create configs as if the build were a postsubmit build. You would
need to set BUILD_NUMBER environment variable now, in order to mock
presubmit configs. Since these configs are for test harness ingestion,
and it is not expected that engineers need to debug any difference
locally, this seems acceptable.

Test: Manual comparisson of treehugger build output.
Bug: 182293323
Change-Id: I7848be417a7d5c0cb46dbf19dbfc67fab938f126


GitOrigin-RevId: e88d14b849464bbcc663c9a2c677917495e5be7b