Introduces support for node reusing to the composer.
When in a reusable group an attempt is made to reuse nodes
when a key changes instead of removing and rebuilding all
the nodes. This has benefits when nodes cache information
about their properties and their children. For example,
this is escpecially helpful for LayoutNode as it caches
layout infomration and if a the tree is substantially
unchanged then the layout information can be preserved
when otherwise it would have to be discarded.

Relnote: """
Introduces ReusableContent which will attempt to reuse the
nodes in its content instead of replacing them when the key
is changed. When the key is changed the previous values in
the slot table for the content is ignored except for the
nodes that were gnerated and the values used to update the
nodes.

Introduces ReusableComposeNode that will reuse the node
emitted instead of replacing it as is done for ComposeNode.
"""

Test: ./gradlew :compose:r:r:tDUT
Change-Id: I1dd86f88d3e96be8aee0a433746f4f3fbc54c3c6


GitOrigin-RevId: 0d24d0677fa7f5492a3d50acee7b64c36c97d379