Lazy list prefetch
Introducing the mechanism for prefetching the next item during the scrolling of LazyColumn/Row. The logic is pretty much the same as the one used in RecyclerView:
After every scroll we calculate what is the next index which would be visible if we continue scrolling the same direction and schedule the prefetching work after the frame. After every frame there could be a gap before the next frame is scheduled to be started. We can use this time to do extra work. In our case we would pre-compose the item which is about to be visible. This makes the frame which will need to introduce the new item with the given index faster as the composition pass is done already.
This change decreases the median frame time of TrivialListScrollBenchmark(95th percentile) by 35%.

Test: tested with a new LazyListPrefetcherTest and manually
Relnote: The next visible item prefetching logic is introduced for LazyColumn/Row
Fixes: 184940225
Change-Id: I8a4bc52322a5444e3782274822fa6d07f7c1d402


GitOrigin-RevId: 4f9e6724f169845b9ce6cb375a23a23b8beab574