Updates ripple related screenshot tests in preparation for migrating ripple to use RippleDrawable
RippleDrawable animates on the RenderThread, so we cannot wait for main thread synchronization. Instead we just wait until after the ripple has finished drawing, so we have a stable screenshot to compare. State layers are still rendered in Compose, so we can fast-forward these as for any other Compose animation.

Also updates MaterialRippleThemeTest to make it easier to see the ripple against the background, and to properly handle unbounded ripples (i.e, by not clipping the ripple and allowing it to draw outside the bounds of the test component).

Bug: b/184842196
Test: tests
Change-Id: I084142379d6e88d61770c291f896ddb35929f493


GitOrigin-RevId: 9a8656d30d34b23386a1843df7baff8c5f3b21c1