Update KSP to alpha10, KCT to 1.4
Update XProcessing to take advantage of the new ORIGIN types:
https://github.com/google/ksp/issues/375

Added support for doc comments in KSP
https://github.com/google/ksp/issues/392

Added support for fun interfaces in KSP
https://github.com/google/ksp/issues/393

Removed custom KCT workaround with the update to KCT 1.4.0

With this update we also get access to the containing class declaration
for compiled top level kotlin functions. That change is more complicated
so added a bug for that as a followup: b/188822146.

We also now have access to a type mapper between java and kotlin classes
but it is not working as we expect so will skip that update for now.
Specifically, it does not support primitives and it also maps java
classes to their immutable counterparts (e.g. java.util.List to
kotlin.collections.List instead of MutableList)

Last but not least, also updated incremental processing test to specify
room version explicitly instead of +. It becomes a problem in playground
where androidx.snapshot repository has a higher version value (SNAPSHOT)
then local publish task.

Bug: 160322705
Bug: 188822146
Test: room tests
Change-Id: I0bc03621bc0727fc94fba58c7c4e2cdbb254424d


GitOrigin-RevId: 5cabf8fcc3a368002c8546047b70b3506733d9ad