Do not reset initial scroll position in LazyColumn/Row when items are loaded asynchronously
If the initial scroll position(or restored scroll position) is not 0 this value will be immediately overridden by 0 if there were no items yet provided for LazyColumn/Row.
It is easily reproducible with Paging where we start loading asynchronously and there is always a first frame when we don't have any items.

Fixes: 177245496
Test: new tests for LazyColumn and LazyRow
Change-Id: I95e374fb06dc480aab7b56b8b3cacb92ca1188a9


GitOrigin-RevId: 16278c75a09262bfc2b1e230730f79ec5a4d4918