Ensure reusing is enabled during composed modifier materialization
During the property update phase of ReusableComposeNode reusing is
disabled to prevent redundant property sets. However, since
materialization is run during this phase it was disabled for
materialization causing materialized lambda values to be incorrectly
reused.

This change enables reuse during skippable update which is where
materialization runs and is the only supported location remember
can be used during update.

Fixes: b/189825174
Test: ./gradlew :compose:ui:ui:cC
Change-Id: I60e7a69bb844515f60abaa0acf8885733dc7d91e


GitOrigin-RevId: 22dde4082d578b030778e97fe43ff40fad86bf2f