Fix offsets for skipping condition and skipToGroupEnd calls.
For something like:

```
32: @Composable fun foo() {
33:   repeated(5) {
34:   }
35: }
```

The current code ends up generating code that looks something like:

```
line 32: ...
  if skipping goto SKIPPING:
line 33:
  repeated()
  ...
  goto END
SKIPPING:
  skipToGroupEnd()
  ...
END:
```

Since line numbers cover all instructions until the next line
number instruction, this means that the `skipToGroupEnd` call
will have line number 33 which is incorrect. Also, it confuses
the debugger so that a breakpoint on line 33 will only be
hit if the body is skipped!

This changes the generated code to look like this:

```
line 32: ...
  if skipping goto SKIPPING
line 33:
  repeated()
  ...
  goto END
SKIPPING:
line 32:
  skipToGroupEnd()
  ...
END
```

Which ensures that there is no line change for the if when
the skipping path is taken. And there is a line change when
the non-skipping path is taken. This also stops confusing
the debugger and breakpoints can be set on the call that
can be skipped.

Test: manual test in debugger
Bug: 190371640

Change-Id: I6ac1f4bd8e5015ec81b3250bc9f748167b769573


GitOrigin-RevId: a8a9fcfa6373e356a1271cdac8cd2d4c80f5a90f