Add contentDescription to Canvas
Canvas is analogous to Image or Icon in that it is inherently opaque to
accessibility and sometimes used decoratively, and sometimes as a
semantically meaningful object.  So it ought to analogously support
optionally providing a contentDescription.  (Because this API change is
coming very late, for backwards compatibility reasons, this CL uses a
slightly different pattern than the nullable mandatory parameter used by
the others.)

In the past, we were OK not providing a contentDescription parameter for
canvas because we felt it was just one representative of a broad
category of use cases that needed custom contentDescription, and we
instead used Canvas as an example of using a `semantics {
contentDescription = }` block in the docs:
https://developer.android.com/jetpack/compose/accessibility#custom-content-descriptions
(related doc change uploaded at http://critique/378219099 )

However, with experience from beta, I've changed my mind.  Because we
prominently mention it in our accessibility docs, we started to
see an antipattern in many sample apps where developers set
contentDescriptions not just on leaf nodes but on high-level entities
like cards/panes, which has had destructive effects on accessibility
usability.  Accessibility team has also recently explained to us that
contentDescription is not intended by ARIA as general-purpose
description parameter but is very specific to things like images, so we
should discourage its use outside of Foundation.

Test: canvas_contentDescription
Relnote: "Canvas now supports a contentDescription parameter for
accessibility."

Change-Id: Ib547c8fdb4c76a64f5ddcfe8ef3d46a0792f40bc


GitOrigin-RevId: 5995a24092b374267c17c7b82d529c3ca22f5cc5