Fix: TextField not always properly visible after keyboard opens
requestRectangleOnScreen may modify the input Rect. When it's called
repeatedly, we may encounter some unexpected issues. This can be one
of the reasons that TextInputService.notifyFocusedRect sometimes doesn't
work.
Before this change,it's very likely that notifyFocusedRect doesn't work
when the keyboard is opened for the first time. After the fix, it no
longer happens in the demo APP.

Bug: 190539358
Test: ./gradlew test
Test: ./gradlew compose:ui:ui-text:connectedAndroidTest
Test: manually tested in demo app
Change-Id: I859a4ab7be2fb0b4bda808646027cf3934af2196


GitOrigin-RevId: 9287a219bf946223ad4119723db9eb2c891aecab