Fix flaky DrawModifierTest.testCacheInvalidatedAfterLayoutDirectionChange()
This test was sometimes failing because there were 3 cache invalidations instead of expected 2. It is failing on some devices, and I expect it was failing because those devices indeed were needed some extra invalidations, maybe because of the Activity resize or something similar. However for exact this test we don't really care how many invalidations happened, we only need to make sure at least one invalidation with the correct layout direction happened. So I updated the test.

Test: run the test
Fixes: 182512695
Change-Id: Icab5ae729154a8d9ca90f910469a36cfcff5d489


GitOrigin-RevId: 2e7475e92113f7cbe694ea84a617c07373637349