Fix: Long-press-and-drag uses wrong selection anchor
Long-press-and-darg selection used the long-press position as the
selection start handle position and the dargged position as the
selection end handle position. This doesn't work with the word based
selection triggered by the long-press very well.

Check the following example:
  Text: abcdef ghijk!
If user long pressed on letter 'c', the word "abcdef" will be selected.
However, if the user starts to drag, and drag to 'e'. Instead of
selecting "abcde", only "cde" is selected.(As explained above, the
long-pressed position is used as the start handle position.)

This CL fixed the issue by only update the end handle position and keep
start handle untouched when the user is dragging.

Bug: 187976146
Bug: 188568899
Test: ./gradlew test
Test: ./gradlew compose:foundation:foundation:connectedAndroidTest
Test: manually tested in demo app
Change-Id: I440c220c68995e1338ccea02ea541471ce60036b


GitOrigin-RevId: 00a8e6557c54646035ae1638da7c1a49c32a9be6