Desktop. Replace java.awt.Image by Painter in Window API
1. Replace java.awt.Image by Painter in Window API
It is better to not expose awt classes in Window API, and keep it platform-inpendent
(in the future we can port Window API to the platforms other than Swing).

This CL also allow us to easely use svg/xml icons in Window titlebar or Tray.

The window icon will be scaled to an appropriate size
Swing internals will request an icon with the needed width/height
which depend on the current display density.

2. Provide a public api for image conversions AWT <-> Compose.
Use the verb `to` instead of `as` because we don't wrap images, we copy them.

3. Refactor getting density: move logic into one place. wrapped.contentScale in skiko
is equal to graphicsConfiguration.defaultTransform.scaleX.toFloat()

Change-Id: I71785119c912a8ef7ff50f744bb5de47db860c16
Test: ./gradlew jvmTest desktopTest -Pandroidx.compose.multiplatformEnabled=true
Test: manual ./gradlew :compose:desktop:desktop:desktop-samples:runWindowApi -Pandroidx.compose.multiplatformEnabled=true


GitOrigin-RevId: f04f0d94fd16dd9ac61ac019fb7a54fb625b3d28