Rewrite touch input injection to facilitate other modalities
`performGesture` will be deprecated, in favor of `performTouchInput`.
Next to that, `performMultiModalInput` has been added to facilitate
multi-modal input. Each modality that will be added in the future will
get its own `performXyz` method with an accompanying receiver scope with
methods that only relate to that modality, and an extra property in
`MultiModalInjectionScope` to access the modality's scope from
`performMultiModalInput`.

The first modality to be added after this CL is mouse input:
`performMouseInput`, `MouseInjectionScope` and
`MultiModalInjectionScope.Mouse`.

There is still only one InputDispatcher, which will offer methods for
all elementary events for each input modality, such as
`enqueueTouchMove` or `enqueueMouseScroll`. If this proves to be
unmaintainable, we can revisit in the future as this is internal API.

`GestureScope` is replaced by `TouchInjectionScope : InjectionScope`.
`InjectionScope` exposes all general properties, such as positional
information of the node. `TouchInjectionScope` contains elementary touch
input methods such as `down`, `move` and `up`. Finally,
`MultiModalInjectionScope` contains a property with the receiver scope for
each modality, such as `val Touch: TouchInjectionScope`, which enables you
to send input from all supported modalities.

Aggregated gestures, such as `click` and `swipe`, have been removed from
the InputDispatcher and are now extension methods on the corresponding
injection scope (e.g. `TouchInjectionScope.click`).

Tests have been moved accordingly. Tests for `GestureScope` have been
kept to ensure backwards compatibility.

Bug: 190493367
Test: ./gradlew compose:ui:ui-test:cC
Relnote: "Introduced `performTouchInput` and `TouchInjectionScope` as a
replacement for `performTouchInput` and `TouchInjectionScope`, paving
the way for other modalities (like mouse).

`TouchInjectionScope` has the same methods as `GestureScope`, with the
exception of `movePointerTo` and `movePointerBy`, which have been
renamed to `updatePointerTo` and `updatePointerBy`. All other methods
are the same.

The behavior of `TouchInjectionScope` is almost identical to
`GestureScope`, with two small details:
1. When sending a down event while pointers had been moved without
   sending a move event (in other words, `updatePointerTo()` has been
   used, but not `move()`, and then `down()` is called), the previous
   implementation would advance the event time and send a move event
   before sending the down event. The new implementation still sends the
   move event, but doesn't advance the event time in this specific
   scenario.
2. When sending an up event while pointers had been moved without
   sending a move event (similar as above), the previous implementation
   would advance the event time and send a move event before sending the
   up event. The new implementation does neither: the new positions of
   the pointers will only be reflected through the up event.

Finally, `TouchInjectionScope` introduces a new method
`currentPosition(pointerId: Int)` to get the current position of the
given pointer."

Change-Id: If11912d1d85da53b0f675f874be29ecde1b8c7cd


GitOrigin-RevId: cfa75245b72f6aaece0252ba082df02099bb6d02