Add title to popup window for screen reader
There is an `accessibilityTitle` parameter that can be set to window, but it's not a public API. So Talkback falls back to the `title` parameter to set the title for the WindowInfo when accessibility title is not provided.
In PopupWindow accessibilityTitle is set to "Pop-up Window". Therefore in Compose we set the `title` of the popup to the same string value.

Fixes: 190080692
Test: manually with Talkback on
Change-Id: Iea4b4a16ea0418bf578f3f8655bae366cf770941


GitOrigin-RevId: 30fcd641567396c07dd3b51f816f238aaccf0b07