Fixes issues with Lint checks in IDE
In the IDE, bytecode elements can be represented by KtLightMethodForDecompiledDeclaration and similar special light representations. These can't be converted nicely to UElements, which was causing some checks to report false positives in IDE, but not in CLI. This CL changes places where we were converting from a PsiElement to a UElement and back, that were causing errors in the IDE - instead we can check the PsiElement directly, which should consistently have the information we need.

Fixes: b/191508358
Fixes: b/193457092
Test: Manual
Change-Id: I90a6d72145c3b356491742b268e77a698f98f38f


GitOrigin-RevId: 92c84c51fc123170896796ed33be297d23946d96