Desktop. Get rid of DesktopTheme, integrate ScrollbarStyle into MaterialTheme/desktopMain
DesktopTheme and DesktopMaterialTheme were added to configure desktop-only scrollbars in material theme.
But they look like it was cut out as piece of MaterialTheme, so it feels right to integrate them into it.

Besides that, users usually didn't understand the purpose of DesktopTheme/DesktopMaterialTheme, and even more often didn't know about them.

So, even if users don't use MaterialTheme, we should provide a non-ugly default style for scrollbars.
Currently we made it simply as in the light material theme.

There are plans to completely redesign of the current scrollbars API, so we can use it on Android too, and make design of scrollbars by the correct material specification (BTW, I haven't found a public specification for them).

We probably split scrollbars into two parts - design-independent (in foundation) and design/platform/input-dependent (in material). How this will be done isn't yet known. Maybe we'll keep the current approach - implementation in foundation, and configuration in material. Or we extract small pieces of scrollbars into small independent functions, and construct a material scrollbar from them in the material module. But that's heavy work, and we can't do it now.

Test: manual ./gradlew :compose:desktop:desktop:desktop-samples:run1 -Pandroidx.compose.multiplatformEnabled=true
Change-Id: Ibb1e5ecc0608f3ee743e382973e30fab8cc09103
Relnote: N/A


GitOrigin-RevId: 4c169c3376fd856988b1b4642c5d2770a7316de7