Desktop. Simple entry point for single-window applications
1. Very often we don't need composable window, custom closing logic, and observing changes in window attributes.
We just need to open the window.

With the current Composable API it is a bit verbose:
```
fun main() = application {
    Window(onCloseRequest = ::exitApplication) {
    }
}
```

This CL adds singleWindowApplication:
```
fun main() = singleWindowApplication {
}
```

Personally, I am tired to write the first variant every time I write a snippet.
But I am biased, because, I write a lot of snippets :)

2. Add a more popular variant of WindowState constructor.

Change-Id: I21cf00fbbf86cb2528563649ed9e253f3c42bcda
Test: manual ./gradlew :compose:desktop:desktop:desktop-samples:run1 -Pandroidx.compose.multiplatformEnabled=true


GitOrigin-RevId: 0477370c20907f11ea7c1295ccb0a14ecf8ea938