Bump androidx.test.core to 1.4.0
Test Core 1.4.0 has some fixes that allows it to work with Android S. We
should update our dependency so that all libraries that depend on test
core will be compatible with Android S.

This was previously reverted in aosp/1773211 due to failures in a number
of large test (i.e.
https://android-build.googleplex.com/builds/tests/view?invocationId=I63900009972892409).

Those failures have been addressed by adding explicit dependencies on
the affected libraries. The failures in camera, compose-material,
compose-ui-inspection, and media2 seem to be unrelated to this change.

RelNote: "Libraries that depend on the `Test Core` library have been
upgraded to version `1.4.0` and will now work with Android platform
version S."
Test: ./gradlew bOS
Bug: 189353863

Change-Id: I88b726b2d76735ac012ea47298e0bf02092cd857


GitOrigin-RevId: fa788954621181fa8131961688634d7f739a354b