Workaround for Lint missing source sets in the MPP build
AGP and MPP don't have any Lint integration, which means that MPP source sets such as commonMain and jvmMain aren't ever added to Lint. There is currently no public API for configuring this in Lint, so for now we use a hack to manually add the source sets to the Lint task inputs. In the future with the new Kotlin Project Model this should not be needed, as there will be better integration from AGP side to configure Lint for multiplatform projects.

Bug: b/195329463
Test: lintDebug
Change-Id: Ibe83a143c382e07d7476abc826ccd967ded1e8b0


GitOrigin-RevId: b3daac3693054c6ff59017b1d9580faf87eba861