Fixes Compose lint tests for Lint 30.1.0-alpha08
Due to https://youtrack.jetbrains.com/issue/KTIJ-18821 annotated local variables do not have annotations on the PsiType - this means that when TestMode.TYPE_ALIAS is enabled there is no way for us to resolve the type-aliased annotation until this bug is fixed. In any case typealiasing @Composable is a strange use case that should probably never happen, so we can just disable this test mode as it is not really relevant for these lint checks.

Fixes: b/196834589
Test: lint tests
Change-Id: If381b199ff4152e5f9592f8154b39818f075d07b


GitOrigin-RevId: 88e61653c0ddb64a9e186d5812469c8117a26627