Update Identity Credential version to 1.0.0-alpha03
With androidx-platform-dev now merged into androidx-main the code for
Identity Credential to use HW-backed Identity Credential on Android 12
is in. Bump the Identity Credential Jetpack version so this can be
released for use in applications.

The change which enabled calling through to HW-backed Identity
Credential originally landed in androidx-platform-dev,
see ag/15027701.

Test: Unit test pass on Pixel 3 XL (both SW and HW-backed)
Bug: None
Change-Id: Ic56ac7a516d015d4aa62352caca6a88cfb7fc2d4


GitOrigin-RevId: 88905f5d7fa8c84418a8b9ef9fb64fb47288ea51