[GH] Declare AndroidXPlaygroundRootPlugin in buildSrc
This PR fixes a bug in playground build where it fails to find
SDK location because it fails to detect the playground plugin.

When a plugin is not declared in the metadata, it does not get
and ID, hence the  check fails. I've added the metadata
and also changed the ID key that we use to search.

Bug: n/a
Test: gw :datastore:datastore-core:generateApi

This is an imported pull request from https://github.com/androidx/androidx/pull/233.


Github-Pr-Head-Sha: f8fbf649bbf96ab421e7a863610505bb881d9c40
GitOrigin-RevId: f1c1de6474d18da629461168558bc7780865b931
Change-Id: I81d34badf1ddf36f558e20c764eb00b64b69d291


GitOrigin-RevId: df9d84a401194996c4effebecdc3d0fd64cf3c27