dismissOnOutsideClick and updateAndroidWindowManagerFlags Popup properties
Relnote: dismissOnOutsideClick was added to PopupProperties, replacing dismissOnClickOutside which was deprecated. The new property receives the click position and the anchor bounds, providing finer control over whether onDismissRequest should be invoked or not. For example, this can be useful to prevent anchor dismissal for touches on the anchor.
Relnote: updateAndroidWindowManagerFlags was added to PopupProperties, offering low-level control over the flags passed by the popup to the Android WindowManager. The parameter of the lambda will be the flags calculated from the PopupProperties values that result in WindowManager flags: e.g. focusable. The result of the lambda will be the final flags which will be passed to the Android WindowManager. By default, updateAndroidWindowManagerFlags will leave the flags calculated from parameters unchanged. This API should be used with caution, only in cases where the popup has very specific behaviour requirements.
Fixes: 193899327
Test: new test in PopupTest
Change-Id: I6e9f9d393dc12bd1ea7ae294af1761817f154735


GitOrigin-RevId: aad79766e7ee607e2ee1689123a6b58c9baa09b5