Update LibraryGroups for Window
Extract window-extensions to its own groupId so that we can update the
version independently. This will be less confusing to OEMs who are the
actual consumers of the extensions.

Extract window-sidecar to its own groupId for similar versions as
extensions.

Merge all of the window module group so that we can use some
additional tooling around internal APIs. One example is using RestrictTo
with a LibraryGroup. This will allow us more flexibility when adding
data classes and infrastructure around overriding behavior for tests.

Move window extensions to beta so that we can use it in core library.
Make all window artifacts have the same library group.
Update the corresponding API files.
Update the gradle files so that all artifacts have the same version.

Relnote: Extract extensions
Bug: 198405125
Test: ./gradlew window:window:cAT
Test: ./gradlew window:window-java:cAT
Test: ./gradlew window:window-rxjava2:cAT
Test: ./gradlew window:window-rxjava3:cAT
Test: ./gradlew window:window-testing:cAT
Test: manual - run samples on a device with window-extensions
Change-Id: I25a5fda9e895d05c05bff1e11f3db6fed4bfa989


GitOrigin-RevId: 29ba5a1849827a83b7f6610b435d6512e9241d5a