Propagate local owners as view tree owners in AndroidView.
Propagates the LocalLifecycleOwner and LocalSavedStateRegistryOwner from
an AndroidView to the managed View by setting ViewTreeLifecycleOwner and
ViewTreeSavedStateRegistryOwner, respectively, on the AndroidViewHolder.
This ensures that, if something in the composition (e.g. a navigation
library like Jetpack Navigation Compose) sets those locals to
owners that are different from the ones used to create the composition
itself, any Android Views nested beneath them will see those owners and
not the ones used by the composition.

Test: Added tests to AndroidViewTest
Fixes: b/179708470
Relnote: "AndroidView now propagates LocalLifecycleOwner and
    LocalSavedStateRegistryOwner to its view via ViewTreeLifecycleOwner
    and ViewTreeSavedStateRegistryOwner."

Change-Id: I38f966576085db66e35b0106f49275ef5bb31adc


GitOrigin-RevId: 11a26697f42c3a6d4a248eb35f52d2c26bd6d9cf