Add flush() and dispose() to InjectionScope
This enables us to abstract the internal details of sending and
disposing of injection scopes to high level generic concepts.

Fix: 200914287
Test: ./gradlew compose:ui:ui-test:cC
Relnote: "Added `flush()` and `dispose()` to `InjectionScope`. Use them
when you want to flush all queued up events immediately and when you
want to dispose of the scope, respectively."

Change-Id: Ifb73a2812da448c872ab86fc06be4f8c242960b3


GitOrigin-RevId: a6ccb47fe76b78e794f175f997d4eb7ba3818c7b