Update CMake version to 3.18.1
This change is required for updating the version of the NDK we depend on (21 -> 23) as well as the version of CMake. The CMake prebuilt no longer comes bundled with Ninja so both those executables have been linked to native-build-tools/bin.

Bug: 196982163
Test: ./gradlew assembleDebug
Change-Id: Icb286b2399bb2f5eae1f4a9a893ee0c33e8aa9a4
Merged-In: Icb286b2399bb2f5eae1f4a9a893ee0c33e8aa9a4


GitOrigin-RevId: 83d23c57fae8b3f82d79ab26a519061f09d6cadd