Fix isolation violation in nested mutable snapshots.
When creating a nested snapshot of a mutable snapshot, the
new snapshot was not created with the correct invalid snapshot
set causing it to see the values of snapshots created after the
parent but before the nested snapshot, breaking isolation.

This change marks all snapshots between the parent snapshot
and the new nested snapshot as invalid in the new snapshot.

This change also fixes `withMutableSnapshot()` to correctly
dispose of the snapshot when it is applied. As this will not
fix all uses of `withMutableSnapshot()` (since the inncorrect
code might be inlined into a library not recompiled with the
change) mutable snapshots now deactive from their parent
snapshot whew dispose even if the nested snapshots are not
disposed.

Fixes: b/200575924
Test: ./gradlew :compose:r:r:tDUT
Change-Id: I0a133558a421a6a9d654e11a22978ad91136b82a


GitOrigin-RevId: 844f9620c04c36a04bcb206ed0938b983acbc85d