Remove focusable modifiers from Progress Indicators
Focusable modifier was added to support a11y on older version of
Talkback, otherwise indicators were not accessibility focusable. The fix
went to Talkback in June 2020.
Focusable provides incorrect UX on desktop as no action really
accossiated with focus in progress indicators.
So instead of making it focusable, we mark progress semantics as merging which will set
`isScreenReaderFocusable`. This will ensure that indicator is not ignored by
Talkback anymore.

Test: manually with Talkback
Test: new tests in AndroidAccessibilityTest
Change-Id: Id956e272baa9306e0c3902424ca95367a6c2b05c


GitOrigin-RevId: 245702cedafa06f9b593afec9159c542b0ce0b49