Do not propagate down in scrollable when fling is caught.
When user flings the scrollable container and then catches it with the fling, we need to make sure we consume down so clicks on the children in the list are not triggered.

Due to migration to suspen input prior to 1.0 this functionaliy has been broken (with no tests to catch this).

This CL fixes the problem by checking and consuming on the initial pass + writing a regression test for it.

Fixes: 202309451
Test: added regression tests + ScrollableTest launched
Change-Id: I319c3b92b7d498e6aad862e0c134beed9afdd245


GitOrigin-RevId: 3ce4be8ba1b1b300ce0ecfd66962c364c3e21ada