Remove flush() and dispose() from InjectionScope
We shouldn't have dispose() on an object which is used as a receiver
scope. Just reaching the end of the block with such a receiver should
probably be disposing it. And flush() isn't really needed.

Bug: 204078232
Test: ./gradlew compose:ui:ui-test:cC
Relnote: "Removed `InjectionScope.flush()` and
`InjectionScope.dispose()`. Flushing of all events and disposing of the
scope now happens at the end of the called perform*Input() method like
before."

Change-Id: I2bed810296ab70173753b06bbc049dabeec39daa


GitOrigin-RevId: 2bd52631d7950e83ae9502836d360d4906981c1f