Improved handling of hit testing in clipped regions
Fixes: 200183104

Relnote: N/A

Previously, if a touch happens outside a clip region, but
within the minimum touch target distance of the clip, the
hit would always transfer to children. This would cause
problems with elements like LazyColumn who want to clip
touch.

Now, a touch only affects children of a clipped region
if that child has a size smaller than the minimum touch
target. Touches outside the clip region now have a lower
priority than touches inside the clip region, so an element
that is visibly touched will be considered "hit" instead of
a touch on an element that is clipped, but has a minimum
touch target that would otherwise allow its touch to escape
the clip area.

Test: manual, new tests
Change-Id: Ib7d8ee85a4da47f383c161a0b7179bc091bda4d0


GitOrigin-RevId: e4f4cac9ba90fe0f0877fb2e2514b09096542972