Handle explicit SAM conversion in ComposableCallChecker
The underlying problem is that the Kotlin compiler discards annotations
when computing the function type for a fun interface. In particular,
this applies to the expected types used during resolution. Since the
ComposableCallChecker needs to know whether the expected type is
composable we have to handle this case separately.

The previous code only handled implicit SAM conversions, by looking at
the enclosing call and checking whether the type of the current value
parameter is a fun interface type. For explicit SAM conversion the
enclosing call is a call to a synthetic fun interface constructor which
expects a function type without annotations.

This commit adds an additional workaround for this case and cleans up
the handling of SAM converted composable lambdas by annotating them in
the frontend.

Bug: 207458633, 207631017
Test: ./gradlew :compose:compiler:compiler-hosted:integration-tests:testDebugUnitTest --tests "androidx.compose.compiler.plugins.kotlin.analysis.ComposableCheckerTests"
Relnote: Disallow composable calls from unannotated crossinline lambdas.
Change-Id: Ie458421cdcfd314fa9bc89e78d9de3213779adae


GitOrigin-RevId: 8102b9c8fb6a52762dbc4f896e858ad960d66297