Remove dead code from ModifiedFocusNode
fun ModifiedFocusNode.requestFocus(propagateFocus: Boolean) has
a propagateFocus parameter that is never used. It is always set
to false. This was added when the code was initially written,
when we thought that we would propagate focus down the hierarchy
to implement default focus. As time progressed, we have decided
that we won't provide a default focus API as users can directly
request focus on an item using a side effect of composition.

This CL removes the propagate focus logic and related tests.
It also enables us to remove a couple of related obsolete TODOs.

Bug: 207040910
Bug: 144126759
Bug: 144126570
Fixes: 207040910
Fixes: 144126759
Fixes: 144126570
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.RequestFocusTest
Change-Id: I352b65a38ec7006cc9abd4abf9f08fa0dc1c1820


GitOrigin-RevId: d3ba314bc11fe091d8e56fc888fc3a639278d980