Support large contentPadding in Lazy lists
Previous approach when the contentPadding was just added via the padding modifier doesn't work if the total padding is larger than the parent size as the size will be coerced to 0. But providing such large paddings is a valid use case and it is supported when used with Column(Modifier.verticalScroll().padding()).
Now we will just apply the content padding manually during the offsets calculations.

Test: Covered new supported use cases in LazyListsContentPaddingTest
Fixes: 199131100
Change-Id: Iba3e1689ee37948667dc2b8a7de06865465ab3f5


GitOrigin-RevId: 0b575945f5276bd1d43c30bcf09629d6df69c9f7