Fix indexOfLastPressed to account for Stylus button reordering
Fixes: 197960375

Because stylus primary and secondary are indexed after other buttons,
they would be considered as last index, even when back or forward
are pressed. However, the stylus primary and secondary indices
are mapped to 0 and 1, and back and forward are 3 and 4. Thus,
the last index should be the back or forward, not stylus primary
or secondary when they are pressed.

This corrects the behavior.

Test: new test case added.
Change-Id: I39013b96bebc118ae79f939e25dddd00d5561a65


GitOrigin-RevId: e03b6e46afba63c708502dad82b2d76c1de3cff8