Allow invalidate() to be called during applyChanges()
Now avoid determining the exact location of an anchor in
order to invalidate it as a writer is active during applyChanges
and using the slot table to determine the location of an anchor
while it is being written is invalid.

This was indirectly triggered by focus changes that are triggered
by change in the focus related modifier of a node.

Fixes: 197064250
Fixes: 197664270
Fixes: 206039948
Fixes: 207800961
Test: ./gradlew :compose:r:r:tDUT
Change-Id: I745fbfc8777beff4e5a78bb15256a247c01974c0


GitOrigin-RevId: 4c48679caa2740853a843ebe8aab9f6dee561e38