Change flans to ToT dependencies
Move Activity, Fragment, and Navigation to use ToT dependencies for
Lifecycle and Activity.

As a result of aosp/1804059 there will now be a dependency conflict
for any modules that rely on compose:ui and activity. Until compose
updates to the ToT Lifecycle dependencies, the modules will need to
resolve the conflict by explicit depending on the proper version.

Test: ./gradlew bOS
Bug: 207012584
Bug: 207012585
Bug: 207012490
Change-Id: I73830d0e00edfb3d917dbbc8f53d9942de98477c


GitOrigin-RevId: 84aea5ac44a2f97a76802e4b84bb3cb8c0c3432b