Add ability to read actual magnifier widget size.
Currently all the magnifier APIs are write-only: properties are passed
down to the `Modifier.magnifier`. Selection magnification needs to read
the size that the magnifier widget ends up being to calculate when to
hide it, but that size is provided by the system so the only way to get
it is to ask the actual widget what it's size is when it's shown.

Test: New UI tests added: ./gradlew :compose:foundation:foundation:cAT
Relnote: "Add onSizeChanged callback to magnifier modifier."
Change-Id: I6879fe3b66f9bfa66abf8401159dc31753f40e9a


GitOrigin-RevId: ae9c81a3fd8d1da284bb13b3eeebe288c685c98b