Correct how records are reused in snapshot state objects
Prior to this change, state objects can get into as state
where the open snapshots can lose the record they they need
to be available because it was incorrectly reused. The
details are in bug 210173523.

This occurred because a state record was consider reusable if
its id was less than the oldest open snapshot and there was
another record that has a greater id that is also less than
the oldest open snapshot. This is insufficient for cases where
the oldest open snapshot has an invalid set as the records in
the invalid set meet the above criteria but should not be
considered candidates for reuse.

This change introduces a pinning table that tracks the oldest
invalid snapshot and changes the reuse to be the oldest open
snapshot or the oldest invalid is that is currently in any
open snapshots invalid set. The table is a double indexed
hash that allows tracking this efficiently (details in the
source).

Relnote: N/A
Fixes: 210173523
Test: ./gradlew :compose:r:r:tDUT
Change-Id: I9281bb0510bcbc6d9d090697468a1dfbf1b5c9a5


GitOrigin-RevId: 3f9fffe7d9f2cf9283d0733816c08537d242e157