Slider: fix the position of the thumb while dragging
Assume that we have a slide with value range = 0..1, left side at 0px,
and the right side at 100px, and it has a track with left side at 10px
and the right side at 90px. Current implementation maps cursor/touch
position 0px to 0.0 and 100px to 1.0, and the thumb does not follow the mouse cursor. The correct behaviour is to map 10px to 0.0 and 90px to
1.0 because it is the range in which the center of the thumb actually
moves.

Test: ./gradlew :compose:ui:ui:clean :compose:ui:ui:test :compose:ui:ui:desktopTest :compose:material:material:connectedDebugAndroidTest, manual testing
Change-Id: Icd8af3e1e9713838df175eae843ef698f04ad8b6


GitOrigin-RevId: 7b559a6b1216a0cb518153d13737a1e7bff00303