Refactoring focus search and fixing some bugs
This cleanup is in preparation to implement a focus cluster.
Focus Cluster will be added as a new foundation API in a separate CL.

When we move focus among children of a deactivated node, we treat the
nodes as if they are children of the first non-deactivated ancestor.
Instead of doing that, we should treat children of a deactivated node
as siblings and move focus among them before visiting other children
of the non-deactivated ancestor.

I also found a bug in tests that use containsExactly(). The assertion
only checks if all the values are present, but does not check if the
values occur in the specified order. We need to use
containsExactly().inOrder() to assert that the items occur in the
expected order. I fixed the incorrect tests and added a test util
function to prevent this kind of error from happening in the future.

Bug: 213508274
Test: /gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.OneDimensionalFocusSearchNextTest
Test: /gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.OneDimensionalFocusSearchPreviousTest
Test: /gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.TwoDimensionalFocusTraversalTest
Change-Id: I68f1d0d78d00567b9719e752fa4b9c1c4ec26386


GitOrigin-RevId: b80d458bc8f842ba493086b84d8a68726c3a3302