Decouple lazy list/grid state from LazyLayoutState
It is possible because each state can provide their own RemeasurementModifier. After decoupling it turned out that most likely there is no real need in allowing state of LazyLayout to be hoisted. Instead lets try to keep it as an implementation detail for now, right now it is only needed for the prefetching.

Relnote: N/A
Test: tests in lazy package
Change-Id: I179a3d34b6a201c86d7b441879a89d5148eb8350


GitOrigin-RevId: 4d34448091f005728301d57f40f65e9c873afbc2