Desktop. Pass PointerButtons and PointerKeyboardModifiers to ComposeScene instead of reading them from AWT event
By default, they are null, and that means that implicit PointerButtons and PointerKeyboardModifiers will be used.
Implicit PointerButtons change in response to event type: if it was Pressed, then primary button state is changed to isPressed.
This is needed for cases, when the calling code doesn't have information about pressed buttons or modifiers, or doesn't want to know (tests for example).

Test: ./gradlew jvmTest desktopTest -Pandroidx.compose.multiplatformEnabled=true
Change-Id: I20be23cb9cf7b5fada06b07770a57e24006e224b


GitOrigin-RevId: e01ac711d466b4988e03387ec9642d1e7c8dad88