TextField action callbacks into state
TextFieldState is being passed to various internal objects; and it
contains the original developer provided onValueChange callback.

However that callback should not be used by any code directly in
CoreTextField or below since it does not enforce the recompose which
ends up stale data in the editing buffer depending on the usage of
the callback.

TextFieldState onValueChange was added in aosp/1447396 as an object
to hold the reference. However it was public and therefore caused
b/200577798 (which is fixed in aosp/1947388.

This CL
- makes the original action callback private in state object
- Adds imeActionPerformed to the state

The parameter can still be used but there is no way
to override or hide it. It shouldn't be used.

Test: Manual test
Test: ./gradlew compose:foundation:foundation:test
Test: ./gradlew compose:foundation:foundation:cAT

Bug: 200577798
Change-Id: I7b410e0c3addf6b24ab28451c32712f1960d06f1


GitOrigin-RevId: bf10468a3974c776fc53ed87fbfe3da9449cde32