Fix observising derivedStateOf objects causing missed invalidations
Observing a dervided state object can cause other invalidations to be
ignored in the same scope if they are sent after the derived state
object is invalidated. Once the invalidation is ignored then the
composition will forget the observation causing subsequent invalidations
to also be ignored until the scope is invalidated for some other reason.

Fixes: 215402574
Test: ./gredlew :compose:r:r:tDUT

Change-Id: I5fb58da9e755a0958b384a7886bac67dc70fda7a


GitOrigin-RevId: ccb2a5e647b2084cbc8dbcb2d04bce43bfb914c4