Report AndroidComposeView's focus rect to the view system.
ViewRootImpl controls panning when the keyboard is shown and the window
is configured to pan instead of resize. When it does this, it finds the
focused view and calls getFocusedRect to find the area to keep above the
keyboard. Because Compose has its own internal focus, it needs to report
that to the view system by overriding getFocusedRect in order to be
moved when the keyboard is shown.

I also factored out the keyboard helper methods from
CoreTextFieldSoftKeyboardTest so I could use them for the scrollable
tests.

Bug: b/190539358
Bug: b/192043120
Test: ./gradlew :compose:f:f:cDAT
Relnote: "TextFields will now be kept above the keyboard when they are
    focused and the keyboard is shown, when the soft input mode is
    `ADJUST_RESIZE`."

Change-Id: I8eaebb684b7828dcf92b0678a86d796b49b349c8


GitOrigin-RevId: 1237c1570308c6a5ad043702f5e9051c1c874602