Refactor focus search to accept a lambda
Refactoring the focus search code so that it runs a lambda
once it finds the next item. This is needed so that we can
request focus on the next item when the next item is beyond
visible bounds. We need this because beyond bounds layout
requests return a block and the items are guaranteed to be
available only within the scope of the block.

For more info, see go/compose-focus-beyondbounds

Bug: 184670295
Test: Internal refactoring, existing moveFocus() tests
Change-Id: I6545ef1a094f5bbe37eb6f861d1ea5ab6a7ec926


GitOrigin-RevId: ef81023fc0de83421e4779f6bef8e14d4ef6a158