Expose parentData on Measured interface
Currently we expose it only on Measurable interface, but it turned out sometimes it is not enough. For example in lazy layouts we created an abstraction which is responsible for providing already measured placeables by index. And then when we operate with a pleaceable we want to know what parentData was provided for it. With also exposing parentData on Measured interface it is now easy to get such data both when you operate with Measureables and with Placeables.

Test: new test in ParentDataModifierTest
Relnote: Measured interface now exposes parentData property
Change-Id: I3313f74c5eb4a75b91cb9d438c6691117b152a7d


GitOrigin-RevId: 5ac267d99394681e368a34c7384accfc478e77c0