Add comments to define Dialog and Popup at a high level.
API council made a request back in the Beta period to document the
difference between Dialog and Popup. The main problem is that Compose's
code comments assumed the developer is already familiar with the UI
jargon and never defines the terms.

The meaning is the same as Android Framework Dialog and Popup, so I
started from the definitions from the DAC pages and elaborated them a
little.  (Overall, the key difference is that Dialog is modal
and Popup is nonmodal, so I used those words in the comments.)

Bug: 176804379
Test: Not needed
Change-Id: I121bcd4b972093ad78ebc7266c67b19e58aeec25


GitOrigin-RevId: 4edab3f39fb9d454a33c8ab74383bd9f7ea68934