Fix races with ImageComposeScene
ImageComposeScene will be introduced in the next commit. It performs in another thread, and have Unconfined coroutine dispatcher. This means all callbacks dispatched to that dispatcher will be performed in the caller thread.

For example, when GlobalSnapshotManager will sendApplyNotifications, all callbacks will be performed in Swing thread, and will race with `render`, which performs in the test thread.

To avoid that, we don't dispatch callback to coroutine dispatcher, instead we use our own dispatching inside ComposeScene via invalidate/render.

Fixes https://github.com/JetBrains/compose-jb/issues/1392

Change-Id: I732b37dd2fa5bb03e7e8f6a001f518199360c9a1
Test: ./gradlew jvmTest desktopTest -Pandroidx.compose.multiplatformEnabled=true


GitOrigin-RevId: bb8c7052845babaf48585c9837df965a2ce756b2