Remove FocusOrder DelegatingLayoutNodeWrapper
Bug: 181981276

FocusOrder now uses a FocusProperties instead of
being its own DelegatingLayoutNodeWrapper subclass

Relnote: "FocusOrder has now been merged into
FocusProperties and focusProperties() now has
all the capabilities of focusOrder(). FocusOrder
and focusOrder() have been deprecated.
focusOrder() that accepts a focusRequester should
be replaced with a focusRequester() modifier in
combination with focusProperties(). This allows
the modifiers to have a stronger separation
of concerns."

Focus order is essentially a property of the
focus node, so it made sense to move the
functionality to FocusProperties. This change
is part of a larger change to remove subclasses
of DelegatingLayoutNodeWrapper.

Test: ran CustomFocusTraversalTest
Change-Id: I601b751755bcfefd849c8a0c0d019e3eaf5d459c


GitOrigin-RevId: 079074340c40b579a9a52df3c9d37f077149cf5c