Nested Scrolling Interop Phase 1: Scrolling compose in cooperating view
In this first implementation of the nested scrolling interop we're adding support
for the case where a Compose scrolling container is placed within a cooperating
view. This means the View can receive scroll deltas from a scrolling child (it already
implements NestedScrollingParent3). This is the case of a CoordinatorLayout.

Test: Added sample and tests to check the behavior.
Relnote: "Enable Nested Scroll interop between View and Compose for
  cooperating View classes. This means compose is now able to dispatch scroll deltas to a (cooperating) View parent."
Fixes: 174348612
Change-Id: I5d1ac5fdeced612ac07f0c26ce14284a43369673


GitOrigin-RevId: fc31a066f4f02a0b9e7a22de99a4255602200a06