Add shape theming support to Material Design 3.
The new shape system is different than it was in M2, and is more comparable to a typescale.

With this mindset, there are two ways for makers to customise the shape system, either by style or by component.

By style: You can customize the shape family and size of any style level in the shape scale. Such changes affect all components mapped to that shape style, except for those with an override.

By component: You can change the style level that a component is mapped to.For example, by default, buttons are mapped to the shape style “full.” If your product requires a smaller amount of roundedness, it can be remapped to the small or medium style in shape scale.

Test: Added shape screenshot test.
Relnote: Added shape parameter to MaterialTheme and Shape sub system.
Change-Id: I374265da432fceb6c72a1b3a61335e7fdefd825e


GitOrigin-RevId: ad9514776247dfac072ef658cd5799a5decab87d