Make Scrollable a focusGroup
This CL makes scrollable a focus group, which ensures that we visit
all the items in the scrollable before moving to the next focus group.

Textfield has a scrollable that is added before the focus modifier.
This causes issues with FocusRequester - The focusRequester is
associated with the focusGroup instead of the focusModifier. This
also causes issues with onFocusChanged observers that would see the
state of the deactivated focus modifier within the focusGroup instead
of the focus state of the focus modifier. I solved these issues by
swapping the order of the scrollable and focusModifier in CoreTextField.

Bug: 184670295
Test: ./gradlew compose:f:f:cC -P android.testInstrumentationRunnerArguments.class=androidx.compose.foundation.ScrollableTest
Change-Id: I149acc64301086f0fd69a99f1e3c816259637367


GitOrigin-RevId: 137c8716674d40eaf9411819e8e5632a7653cb08