Disable LayoutCoordinatesHelperTest.onPlaced_invocation
Disable the test until it is fixed as it is blocking presubmit.

Also improves the assertions in the failing test to compare all values
in the invocations array instead of just the first one that fails, which
makes the failure a lot more usefull (did all values fail or just one?).

And while I'm at it, remove unused @OptIns for experimental API that's
no longer experimental.

Bug: 225198728
Test: ./gradlew compose:ui:ui-test:cC
Change-Id: Id30602076d280eb08a2d9b5a6c77b13fc2254bfa


GitOrigin-RevId: 68cae630f37fe08d85a8ed7b75da77687ad735ae