Nested Scrolling Interop Phase 2: A scrolling view child inside a compose parent
In this phase I made AndroidViewHolder to implement NestedScrollingParent3.
That way it will act as a nested scroll parent for a containing view.
At the same time, it uses an internal NestedScrollDispatcher to forward
deltas to a Compose Parent, so it also can act as a nested scroll child
for a compose parent.

Test: Added sample and tests to check the behavior.
Relnote: "Enable Nested Scroll interop between Compose and View in the
direction Compose > View. This means that a compose parent will be able to
receive nested scroll deltas from a nested scroll view."

Fixes: 174348612
Change-Id: If7949ccb50ba7897bc740b489fd3f4d615a5369b


GitOrigin-RevId: 697235331591da8efbc20a8f4a9896b2e17b0c1e