Refactor TextStyle: delegate to Paragraph/SpanStyle
TextStyle had its own properties that are also properties of SpanStyle
and ParagraphStyle.

Instead of keeping a separate set of values internally, TextStyle only
holds references to ParagraphStyle and SpanStyle, and it delegates the
queries to those classes.

This was also previously suggested in a code comment.

In order to do that
* TextStyle default constructor is changed to accept ParagraphStyle
  and SpanStyle
* Property getters are defined on TextStyle

Test: ./gradlew compose:ui:ui-text:test
Test: ./gradlew compose:ui:ui-text:cAT
Test: ./gradlew compose:foundation:foundation:test

Bug: N/A
Change-Id: I5502bad29b1a15fc8b40d7512df3d201b3ebbea7


GitOrigin-RevId: 0ac70680646535f2cc2f6896e11e869a6ae1ec9e