[Slider] Fix issues when thumbs overlap and raw offsets change values on next gesture
When dragging thumbs over each other raw offset values will be outside the max/min
value for that thumb. If in the next gesture we drag the other thumb those min max values
change causing two thumbs to update at once.

To avoid this, reset the rawOffset to the actual value after dragging the opposite thumb.

Test: Added unit test

Bug: 227344252

Change-Id: I6deac136274482d36eeef43077fa1e7749aaf14f


GitOrigin-RevId: 3fcccb0429f645706caa2576ea8bcd1fdac832a7