Fix flaky testWebViewIsRelaidOut_afterPageLoad
Bug: 227616700

WebViewClient can be called asynchronously from the UI thread,
so synchronization between draw calls and the client can be
flaky. This change waits until the final WebViewClient call
is made instead of having to synchronize at the expense of
testing on API 21 & 22 devices.

I couldn't verify that this fixes the test. I couldn't
make the test fail without this change.

Test: ran test multiple times on cuttlefish
Change-Id: Ie9c2862b91147cb4637559f2a15fa84861372169


GitOrigin-RevId: d720c027abad6e368c9be32858073c6a4491a5f5