Line height behavior and vertical alignment
This CL adds LineHeightBehavior to the TextStyle and
ParagraphStyle. LineHeightBehavior controls whether line height is
applied to the top of the first line and to the bottom of the last
line. It also defines the alignment of line in the space provided
by TextStyle(lineHeight).

go/compose-lineheight-distribution

RelNote: "Added LineHeightBehavior to the TextStyle and
ParagraphStyle. LineHeightBehavior controls whether line height is
applied to the top of the first line and to the bottom of the last
line. It also defines the alignment of line in the space provided
by TextStyle(lineHeight).

For example it is possible to get a behavior similar to what CSS
defines via LineHeightBehavior(
alignment = LineVerticalAlignment.Center,
trimFirstLineTop=false,
trimLastLineBottom = false).

trimFirstLineTop, trimLastLineBottom configurations works correctly
only when includeFontPadding is false."

Test: Added demo
Test: Added tests
Test: ./gradlew text:text:test
Test: ./gradlew text:text:cAT
Test: ./gradlew compose:ui:ui-text:test
Test: ./gradlew compose:ui:ui-text:cAT
Test: ./gradlew compose:foundation:foundation:test
Test: ./gradlew compose:foundation:foundation:cAT

Bug: 181155707
Change-Id: I973329a540ca9f5a6e225f1e5aaffeaa1ff9cc61


GitOrigin-RevId: d0853b60893256b272ed9b3bd444b209de707118