Keep more item as reusable in Lazy lists and grids
After aosp/2056467 was merged there are no bad side effects from keeping items in a reusable pool as we do not keep them composed anymore. This allows us to safely increment the amount of items we keep. It should improve the performance in cases when fast scrolling is happening on the slow devices. I decided to start with the same number of items as RecyclerView keeps by default as it is the time-tested constant used there for years.

Test: updated tests to align with the new constant
Change-Id: I280c1177162980068e9b2b22aacebfba1f47d352


GitOrigin-RevId: 8b69435f88b1ce9544f9d7b230473304fa46cf42