Focus search should only consider placed items
Focus search should ignore items that are not placed.
Sometimes, we compose and measure items but don't place
them. These should be excluded from focus search. A
good example of this is when LazyLists reuse layout
nodes. The cached layout nodes are attached to the
hierarchy but should be ignored by focus search if
they are not currently in use.

Bug: 184670295
Test: ./gradlew compose:ui:ui:cC -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.FocusSearchNonPlacedItemsTest
Change-Id: I995990df2a171bb08cc1be2ed1d1db35cd65027a


GitOrigin-RevId: 7979439ccc7bed82fffa252aa4794f37dc8e565d