Extend VerifyPlaygroundGradleConfigurationTask to check all properties
This change is in response to a build breakage due to missing the gradle
property, `android.experimental.lint.missingBaselineIsEmptyBaseline=true`.
We previously only checked for values of playground properties that were
also declared in aosp due to playground needing a different subset of
those properties. We now use an allowlist to make that subset explicit
to avoid future breakages.

- Added an allowlist to ignore properties which we know to differ
  between playground and aosp setups

Test: ./gradlew verifyPlaygroundGradleConfiguration
Change-Id: I8652e788954e8dd2577f0ac0a8816b7f320020a3


GitOrigin-RevId: 3fd7b2f5c3960368f5680d14891e377322199645