Improve BringIntoViewResponder dispatching behavior for concurrent requests.
The initial fix for this, aosp/2065910, did not ensure that
BringIntoViewRequester.bringIntoView didn't return until the request was
satisfied or interrupted, and couldn't queue requests.

This change queues all overlapping requests, ensures bringIntoView
doesn't return until the previous request is cancelled or complete, and
adds a bunch of tests for various overlapping and interrupting
behaviors. It also improves the documentation a bit.

Fixes: b/216790855
Test: New tests in BringIntoViewResponderTest
Test: ./gradlew :compose:f:f:cDAT
Relnote: "`BringIntoViewRequester.bringIntoView` will now always suspend
    until the request is either completed or was interrupted by a newer,
    non-overlapping request. Overlapping requests will be queued."
Change-Id: I43e7f73af86615f08e2d2d05bfd05ac96d0c65e4


GitOrigin-RevId: c1cf16ad7e547a8e267fd25fb6698ecbbb0ca58a