Add material3-lint and ScaffoldPaddingDetector
This CL adds the material3-lint library (automatically included in
material3) and the first rule, ScaffoldPaddingDetector for when the
inner padding parameter for a Scaffold isn't used.

This is mostly a copy of
https://android-review.googlesource.com/c/platform/frameworks/support/+/2017215
with changes to be for material3 and adding the new module.

Relnote: "Add lint check to material3/Scaffold to ensure that the inner
padding is used"
Test: New tests in :compose:material3:material3:lint

Fixes: 226951418
Change-Id: I72293ba593b57caaacee5d86702b10b281b0246c


GitOrigin-RevId: 2930942005e8fe176ffdda870f2c73debf1461ab