Send apply notifications after coroutine resumes
Add a ContinuationInterceptor to the CoroutineContext passed to the
Recomposer that calls `Snapshot.sendApplyNotifications()` after each
continuation's resume.

This means that state changes made in coroutines that are launched in
composition will be applied immediately after the coroutine has run,
rather than relying on the GlobalSnapshotManager to apply them the next
time it gets hold of the main thread. This will make a difference when
the test coroutine dispatcher is busy dispatching from the main thread
itself.

Fix: 196060592
Fix: 219467661
Test: ./gradlew bOS
Change-Id: Id4cecc74395a9b9fa7bbcffa173c15b73350bb76


GitOrigin-RevId: 445a0d820da9d8b3ba9644a4f3b85c25ae6a0a13