Fix flaky quickAnimation test
Fixes: 230781717

This CL fixes some timing issues in the quickAnimation test and
also some timing issues in the tested method. On R, an animation
can be canceled and the value received in onApplyWindowInsets()
immediately following the onPrepare() call should be used. This
CL chooses to use the onApplyWindowInsets() value if there has
not been an onStart()  or onEnd() call before the posted Runnable
is executed.

Test: ran flaky test 50 times on R and S emulators.
Change-Id: I52dc3396d7629c40025ff4b253c3ef4fc35ffd05


GitOrigin-RevId: 8dd6fc5c2a0811d8eb5c2213cfef008c733c8363