Screenshot Test for Popups - Composite Bitmap
In this CL I use UiAutomation to capture a bitmap of the screen and underlying components.
PopUps live in a different window and compose hierarchy than their containing components, therefore we need
to yse WindowInspector to retrieve all windows in the screen and search for the PopUp one to retrieve its bounds.

Test: Modified existing tests and added new ones to check the behavior.

Fixes: 230837971
Bug: 208991956
Change-Id: I15bd1bdc748e5f5f7bfcf94c45c1bb425fa93cf3


GitOrigin-RevId: e7f58feea62bdc557370d2187318a44542d0f63a