Fix KMP source metadata (reland)
This should change things so that ./gradlew :docs-public:unzipSourcesForDackka
completes successfully when we include the sources for a kmp library.

This is a minimal rework of aosp/2118675, which caused b/236380911 due to a
bug in publication registering.  Here is the small diff with that approved CL
(I also finished removing an unused configuration variable and a stale TODO):

https://paste.googleplex.com/6596120005312512

Bug: 233055013
Test: ./gradlew :buildSrc-tests:test --tests androidx.build.*
Test: http://go/saffgit/scratch/+/refs/heads/main/docs-test/test.sh
Test: ./gradlew \
    --ci \
    saveSystemStats \
    test \
    -Pandroidx.ignoreTestFailures \
    -Pandroidx.displayTestOutput=false
Test: ./gradlew -Pandroidx.compose.multiplatformEnabled=true \
    :internal-testutils-kmp:tasks
Test: http://go/saffppr/count-jars.sh
(Should show 386 published jars at time of commit)
Test: http://go/saffppr/check-tiles.sh

Change-Id: Ie684d08d6dbd871084404af6112f504bb99c76cd


GitOrigin-RevId: 5d98fb922afd9df8c2769e456bd38e2f124c424c