Use latest version of runtime-saveable in ui and ui-text
Modules runtime and runtime-saveable are in the same group and as those modules uses the latest runtime they should also use the latest version of runtime-saveable.
We do not expect our users to add this dependency manually, and not using the latest version here means our users will not get the latest bug fixes we provide in there.

Test: n/a
Change-Id: I8839e32b740584ac8fd1acbe64c052e95f4477dc


GitOrigin-RevId: 9db213a23c3faca8cd2b9440846891cda2d4f63c