Add lock key state getters to key injection API
Adds getter methods for determining the states of lock keys to
`KeyInjectionScope`. Also adds a function that enables injecting
the same key repeatedly a given number of times.

Additionally, adds code to pass the meta and lock key state of the
keyboard to the mouse injection functions, to facilitate multi-modal
input such as ctrl+click.

Bug: 229831515
Test: ./gradlew compose:ui:ui-test:test
      && ./gradlew compose:ui:ui-test:cC
Relnote: "Added `pressKeyTimes` as well as `isCapsLockOn` and friends
to `KeyInjectionScope`. Additionally, the API now supports mouse and
keyboard combined injection patterns such as clicking a mouse button
with a meta key held down."

Change-Id: I4c8dae112eeb7526cbc32b9bce90d7e3bb5ce73a


GitOrigin-RevId: c89f385724e8b3177b2ace26dd35b59ed1fb6652