Offline KMP builds
This CL ensures KMP native builds run offline. Actual build files
will come in a followup but you can see a sample build here (with nsjail):
https://android-build.googleplex.com/builds/abtd/run/L61700000954898119

Kotlin needs 2 set of downloads to compile native:
* Compiler toolchain
  This is based on the host machine and includes llvm compiler
* Target toolchain
  This is based on the compilation target and includes native dependencies
  (e.g. sysroot libs etc).

There are 2 actions we take to ensure Kotlin compiler does not download
anything from the internet:

* AndroidXImplPlugin.overrideKotlinNativeCompilerRepository
As of 1.6.21, there is no way to set URL to download compiler toolchain.
Kotlin 1.7 will have that feature (nativeBaseDownloadUrl property).
The code that downloads this dependency does so by creating an ivy
repository based on the host and resolving it.
Our workaround hooks into the creation of that ivy repository and
replaces its url with the konan prebuilts directory.

* -Xoverride-konan-properties=dependenciesUrl
This property is passed into KotlinNativeCompile tasks so that it
downloads the target dependencies from prebuilts/androidx/konan

Bug: n/a
Test: ANDROIDX_PROJECTS=KMP ./gradlew bOS

Change-Id: I5102ece4ba7dc10fb9483983b72085eeb4602fe3


GitOrigin-RevId: 6210e2b43c6fa6ea5265515f132fc2aec1ed461c