Fix scrollable focused child tracking.
When calculating whether we need to scroll to keep the focused child
in-view after the scrollable viewport is resized, we were using the
bounds of the focused child to figure out if it was previously visible.
That was incorrect, because if the viewport is animating while the
scroll animation is running, it's possible for the viewport to shrink
fast enough that the focused child isn't actually visible, but would
have eventually became visible after the animation, so we need to
continue to animate.

This approach is documented as Option 0 in the design doc:
https://docs.google.com/document/d/16yYJOIcde0Pm7vB--ExSdrY7lrZQ2rqHIn6nYWb7ccg/edit#heading=h.ytkawqefedyn

Note that while this fixes b/230756508 and will cause the animations to
end up in the correct final state, it makes b/220119990 more visible,
since it looks like the scroll animation doesn't start at all until the
viewport animation stops.

Fixes: b/230756508
Test: Manually using the 'Foundation > Scrollable with focused child"
    and "Text > Input fields > Inside scrollable" demos."
Test: ./gradlew :compose:f:f:cDAT
Relnote: "When a scrollable has a focused child, it will now correctly
    scroll to keep the focused child in view when its size is decreased,
    even when the size is animated."
Change-Id: I8086717b14174e566b299f2643f1dd6c0b250773


GitOrigin-RevId: cb90e2addc50ff6f4e7f48df0dc0a03343898ccb