Update Dackka build task to pass in JSON metadata to Dackka
This is the initial integration with a placeholder json file,
SampleLibraryMetadata.json, which is taken from Dackka test data:

https://cs.android.com/androidx/platform/tools/dokka-devsite-plugin/+/master:testData/SampleLibraryMetadata.json

Displaying the generated data is controlled by the feature flag
`showLibraryMetadata` in `AndroidXDocsImplPlugin.kt`, which is set to
false.  Thus, merging this CL will not change the output of the
generated refdocs.

Tests:
1) Run the `dackkaDocs` task before and after this change, and verify
   that the generated docs do not have added metadata
2) Change `showLibraryMetadata` to true and run `dackkaDocs`. Verify
   that the generated docs have added metadata.

Test: presubmit
Bug: 239076511
Change-Id: I39fc595fb2b4c62da64ec6e47e9567f45a6ab09f


GitOrigin-RevId: c6c4febcf22aeddc3a4b8a8e3ab77a5ac94852f4