Validate the indices returned by OffsetMappings and throw better exceptions.
Instances of OffsetMapping are provided by developers when passing a
custom VisualTransformation to a text field. Those offset mappings must
correctly map indices between the original and transformed strings.
However, it is very easy to write a mapping that returns invalid indices
that are outside of the actual strings. When that happens, the text
internals eventually throw an exception, but it's not very helpful. This
change explicitly validates the returned values on every call, and if
they're invalid, throws an exception with a more detailed error message
that can hopefully help developers troubleshoot what's wrong.

The validation is performed by wrapping the OffsetMapping in one that
knows the lengths of the original and transformed text, and then using
that wrapper everywhere we perform the mapping.

This change adds UI tests that validate the detailed exception is being
thrown when interacting with cursor and selection in a text field that
has an invalid mapping.

Bug: b/229378536
Test: ./gradlew :compose:f:f:test
Relnote: "Text fields will now throw more detailed exceptions when the
    `OffsetMapping` provided by a `VisualTransformation` returns invalid
    indices."
Change-Id: Ie73f9a652c29dde66d073a868a2e1c181feaa98d


GitOrigin-RevId: 2858b9e78538908fc3ce28ad346321fd014ae55b