Update state in TextInputService with plausible one immediately
Some IME requests surrounding text just after setting composing
text or commit text. However, Jetpack Compose ask developers
when IME set composing text or commit text via onValueChanged
callback and update the state later. Because of this design,
immediate surrounding text request replied with old state, then
IME thought the previous API was failed and retry or recover
the input session.

To avoid this issue, update the state in TextInputService with
plausible one immediately and reply with that result to the IME.
When the real state arrives, this plausible state is overrided
with the real state. If the plausible state is different from
the real state, i.e. developer modified the buffer, we already
send a selection changed callback to IMEs.

Bug: 239471016
Test: locally verified with Gboard (JA, US), ATOK and SwiftKey(JA)
Change-Id: I02658a8d861fd7cf86ac0dde25548c51ba088e7a


GitOrigin-RevId: b95da1b37888777a035e27fca6efe86ce0e9bb17