Parse AndroidX artifacts to generate metadata entries
This is the next CL in a series towards displaying artifact ID in reference docs. Note that the flag for showing the metadata (`showLibraryMetadata`) is still set to false, so this feature is a no-op in terms of generated refdocs.

This CL has a few items that will be addressed in future CLs:

- Programatically generating `sourceDir`
- Refactoring `MetadataEntry`

Bug: 240309530
Test: Run `./gradlew  :docs-public:generateMetadata` and verify the contents in the generated JSON file (in out/androidx/docs-public/build)
Test: presubmit

Change-Id: I10c1e9f1e392c3388e533dcba800efbe4f6f45ac


GitOrigin-RevId: 955ba228c264e2a22f64afe8339c656a3164aa5d