Simplify SemanticsSort tiebreaking logic
This deletes the tiebreaking-by-size logic in SemanticsSort.  Rationale:

- After aosp/1990935, Material3 Scaffold a11y ordering broke (contents
focused before top bar).  The reason is that to account for
possible transparent top bars, the contents now start at (0, 0).  The
content is taller than the top bar so it always immediately won the
semantics sort.

- If we simply delete this tiebreak, then the logic continues to decide
based on children, followed by (implicitly) z-order and placement.
These are principled strategies that fix this top-bar use case.

- I don't know of a use case solved by the size tiebreak.  All the cases
I can imagine are better solved by children or z-order or placement.
The only reason this code is here is because it was present in the
ViewGroup.java code this logic imitates.  My theory is that it
never did anything either useful or harmful in View world because the
situation didn't come up, but in Compose it proved harmful.

Bug: 232928023
Test: testChildrenSortedByBounds_sameOffset_differentSize
Change-Id: I48e24f42a6c9234a128a877fcdd5b62c03caed35


GitOrigin-RevId: 9779f578a06c8d929bae3be871faa6bb8a155bab