Update the selection in BasicTextField String overload to be consistent with the previous state.
BasicTextField String overload manages its own internal TextFieldValue state to be used by CoreTextField. This state is not directly passed to CoreTextField though, it is copied using the latest String value. If a selection is active and a text that is shorter than selection is passed to BasicTextField, selection coerces but this is not stored in the internal state. This change introduces a SideEffect that is going to run at the end of composition to make sure that a new selection value is stored in the internal state.

Fix: b/239965332
Test: :foundation:foundation:cAT
Change-Id: Ib479d860a453db7bb82c734ec836bc2da06091f3


GitOrigin-RevId: c90f566842b4c47b947f3c3ff0fc818b2686e972