Fix for lazy layouts prefetching not being canceled when we snap to other position
If we schedule a prefetching and then snap to other position straight away it was causing prefetching to proceed with composing an item which is not needed anymore. It was happening issues when we do teleporting during the animated scrolling and when Paging library is used as this composition for the item which is not even located near the currently visible items is causing to do start preloading incorrect next pages.

Test: new tests for lists and grids
Change-Id: I26d1fa6c106d6f95f9e8d912c59eea3f36e251bc


GitOrigin-RevId: 7425fe9a20e9f40dbbeff00efc2cc3212988a6a4