Prepare BringIntoView internals to be able to handle changing requests.
This change modifies the signature of internal bringIntoView-related
calls to take a function returning a Rect instead of a single Rect, and
updates the bringIntoView modifier to calculate the default size of the
request lazily when this function is called, which is required to
correctly animate nodes that change size during the animation
(b/241591211).

This change does NOT actually fix the implementation in Scrollable to
correctly handle changing bounds, that will be done in a follow-up.

It also does NOT change the main public API for bringIntoView, the
BringIntoViewRequester.bringIntoView function – that still takes a
single Rect.

Bug: b/241591211
Test: ./gradlew :compose:f:f:cDAT
Relnote: "`BringIntoViewResponder`s are now able to get the most
    up-to-date bounds of a request while processing it."

Change-Id: If86a581463c14db3fa5638720aabca9830720630


GitOrigin-RevId: b5b29bbc4bef0f1a172742809f0cc00c7ec48391