Revert new public API EditCommand.toStringForLog()
This change was rejected by the API council. It was originally
introduced in
https://android-review.googlesource.com/c/platform/frameworks/support/+/2151717/.

The downside of this approach is that, because EditCommand is an open
interface, external implementations can't define their own log string.
But it's unlikely there are any external implementations of this
low-level interface anyway, so it's not a big deal.

Fixes: b/241785036
Test: Existing test coverage.
Relnote: "Removed recently-introduced `EditCommand.toStringForLog()`
    method."
Change-Id: I8c51930bb9a6e98fe79d24a8e4f09d1ce34588d7


GitOrigin-RevId: 69d2c9a42c80284ec040678b4b544ac846fbd3d4