Allow movable content to move into and out of SubcomposeLayout
Modified how movable content is tracked to allow movable content
to move into newly created SubcomposeLayout and out of
SubcomposeLayout that is being disposed. This is partically useful
when movable content is moving in an out of BoxWithConstraings that
is created and disposed as the content is moving.

The lifetime of when movable content is available to be moved was
extended from just after all composition completes to the end of
all scheduled work for the frame (just before the Recompose waits
for new work).

Fixes: 235398298
Test: ./gradlew :compose:r:r:tDUT
Change-Id: I1558232ce2133c5e55ca5065fc0e76f31d61b37c


GitOrigin-RevId: 07b6f4372eb02255568de41924515ec50d0acd94