Fix crash with Android Views and Talkback
Fixes: 240139002

When Talkback is active, removing a view can lead to drawing
immediately after the view is removed. The internal structure
isn't ready for a draw in the middle of applying the composition.

This delays removing the Views until after the changes are
applied.

I tried to create a test to reproduce the bug without Talkback
active, but was unsuccessful. I'll leave this as testing
only manually.

Test: manual testing
Change-Id: I8e70b738d7e3d29d87c5ab3e3defff8473de1ddc


GitOrigin-RevId: 53223f0ecf7af90e93c39ee1dfd36e238235d888