Add ability to generate unbundled docs for KMP projects
This CL adds 2 new tasks that can be used to generate offline docs
using dokka.

Each AndroidX project has a dokkaPartialDocs artifact if it has the
Kotlin plugin installed. (built by DokkaPartialDocsTask and then
zipped)

There is also a new "kmp-docs" project which declares a dependency on
projects that will be documented.
Documentation task setup is done by the AndroidXKmpDocsPlugin.
This plugin uses the DokkaCliCombinedDocsTask to produce the final
combined documentation.

Output is deployed here:
http://androidx.github.io/kmp-eap-docs

Bug: 249096220
Test: local

Change-Id: Ib9d30df45301b934d7b2cb1e9687c95a7cba5bad


GitOrigin-RevId: 637b5f8d982d0820ac8a8c9b65ecf8fce39ace44