Fix performance issue in BeyondBoundsLayout
Currently when BeyondBoundsLayout is called for a direction opposite to the
orientation of the LazyList, we still add beyondBoundsItems. This causes
a performance issue for focus search where we have to compose all the items
before we exit a focus group. This CL fixes that issue by not adding any
items if BeyondBoundsLayout is called with a direction opposite to the
orientation of the LazyList.

Bug: 249990124
Fixes: 249990124
Test: ./gradlew compose:foundation:foundation:cC -P android.testInstrumentationRunnerArguments.class=androidx.compose.foundation.lazy.list.LazyListBeyondBoundsTest
Change-Id: I723a3a287573b831a0e3a3d2e33edd04aadd7c01


GitOrigin-RevId: e9966146a1ecca76f52dbf58366177dc145ca774