Send apply notifications after Recomposer finishes frame.
This fixes a bug where updating the intrinsics changes would be delayed
a frame. Changes to the intrinsics measure policy in a
layout node during apply changes was not applied until _after_ the
layout pass, which might result in not invalidating layout that should
have been that frame. And since those changes would only be applied for
the next frame, it would invalidate that frame unnecessarily. In other
words, intrinsics changes would be delayed a frame.

This is a prerequisite for b/222093277 (aosp/2243539).

Test: RemeasureWithIntrinsicsRealClockTest (the current test does not
    catch this bug because of b/222093277).
Test: RecomposerTests
Relnote: "Snapshot apply notifications are now sent after the
    `Recomposer` finishes applying changes."
Change-Id: Iad6c0dcd163a5a8f9c5aec426da3d4f701ca509f


GitOrigin-RevId: 98f2641f492c7ffc91e77e39d60b77909669f0ea