Don't restore ModalBottomSheetState if initialValue changed
If the initialValue changed in the composition after a configuration change, we'd still attempt to restore the previous state. If the `currentValue` (saved by ModalBottomSheetState.Saver) was `HalfExpanded` before, this would result in a crash when rotating from portrait -> landscape as there is no HalfExpanded state in landscape.

Test: modalBottomSheet_tallSheet_rotateLandscape_openSheet_staysOpen
Relnote: Fixed a bug where ModalBottomSheetLayout would crash if it was HalfExpanded when rotating from portrait to landscape. Please ensure you are passing in the correct `initialValue`, for example by checking the configuration.
Bug: b/182882364
Change-Id: Ie8df79fae22ea27329024e0ea3b3f6a240151b87


GitOrigin-RevId: faf197e56d6c5af3071dc9e4749b83a8dc181b95