Pass previousTargetValue to AnchorChangeHandler
Previously, accessing the targetValue from AnchorChangeHandler#onAnchorsChanged meant it would be re-calculated with the new anchors, thus possibly not reflecting the actual previous target value.

This fixes a bug in ModalBottomSheet.kt's AnchorChangeHandler implementation where sometimes value closest to the current offset from the new set of anchors would be animated to when changing the anchors, instead of using the actual previous value.

Test: swipeable_anchorChangeHandler_invokedWithPreviousTarget, modalBottomSheet_shortSheet_sizeChanges_snapsToNewTarget
Bug: b/201628457, b/230652481
Change-Id: I0aef54a85527ba204d60ef88cb6bcf2c20440ef2


GitOrigin-RevId: c2f2c64e021479a8c21fbb3e1d2354d79d13ffc9