Refactor the M3 CalendarModel
Rafactor the model to have basic support for desktop and to be able to
call it from the Material 3 lib without a build failure.
The change splits the CalendarModel implementation using `expect` and `actual`, and relocate the legacy implemenation to the common Material 3 location. The API 26+ implementation was left under `androidMain`.

Test: CalendarModelTest
Bug: 245821760
Relnote: N/A
Change-Id: I0dd8b3f738be536cf0bd527ac037036abe295f6e


GitOrigin-RevId: d97118da4d253a311fe90c8e2d709d40ec532e90