Introduce marquee modifier.
Unlike my first attempt at this, aosp/1931284, this time I tried to copy
the simplicity and behavior of the platform behavior as closely as
possible, with the simplest implementation possible.

However, like with that first attempt, this modifier can be used on
anything, not just text.

The modifier uses defaults that match the behavior of the platform, with
limited customization options for repitition count, speed, delay, and
spacing. This should solve all of the use cases of the platform feature,
and if we need to build something more complex we can do that later.

This implementation does not support content that accepts pointer events
or other input. I checked the platform implementation and it also
doesn't handle this well (e.g. if the text has a clickable link in it,
the link is only clickable at certain points in the animation and not
where it's rendered).

Fixes: b/139321650
Test: BasicMarqueeTest
Relnote: "Introduce experimental `Modifier.basicMarquee()` for displaying
    content with a scrolling marquee effect."
Change-Id: I2df44c3343afa8400b0da768a642b77da94c103d


GitOrigin-RevId: 3e74e5ef9d8e4ab6c599c95fcc510fae230c2470