Add Compose tree stitching to ComposeViewAdapter
Different trees from multiple compositions need to be attached as a
single tree for the code navigation structure to make sense when using
elements like LazyColumn or LazyRow. These might end up generating
multiple separate tree that need to be unified as one.
To unify the trees, we use the LayoutInfo that is part of the tree to
find how the different roots are related, merging them into one single
tree.

Test: ComposeViewAdapterTest
Change-Id: I5d108175d234c18bd54ba8f930236c83faa2c133
Relnote: N/A


GitOrigin-RevId: 0da65687b42f12b55b3bcfbb736e8aea625f841d