Improve memory reference characteristics of Snapshot state
Will now overwrite the state of state records that are reusable when
the global snapshot advances and when nested snapshots are applied. This
reduces the number of dangling references to objects that are not
selectable by any snapshot, allowing these references to potentially
be collected.

This change does not detect all such dangling references. A more
exchaustive technique will be added in a follow-on change.

Bug: 203102179
Test: ./gradlew :compose:r:r:tDUT

Change-Id: I04ff2e1b8766615590908bafb79fda135744c506


GitOrigin-RevId: 93fcae828b72670cee2f2388877ba15805dfb451