Add in IsContainer property on Surfaces
This is part of a larger change to TalkBack focus order in Compose that will use the semantic significance of elements such as surfaces in determining traversal order.

Test: SemanticsTests.containerProperty()
Change-Id: I63379fde102abbee7d5464c50b1c86a59e01e768
Relnote: Added in IsContainer semantics property on Surfaces. This property will be used in a later change that determines traversal order based on the semantic meaning of elements such as surfaces.


GitOrigin-RevId: 5361cd861c20ba73c4db0757ef1772574dd593ec