Use correct key for non-reusable nodes groups
Non-reusable nodes use different group key when replaced (125 normally, 126 when replaced). The node was hardcoded to use 125 as a group key in the writer, causing compoundHash miscalculation.

This change updates reusable node to always use the same key (125), while introducing new type of group (ReusableNode), which is ignored when composition is on a reusing pass.

Fixes: 263760668
Test: Updated runtime tests

Change-Id: I8e6ce411e5cd189f7582e8ad6eac76a820569e4a


GitOrigin-RevId: 6388d8da9771f3921f261e9bcf9c76153c7584e4