Fix selection of text that has maxLines configured.
If a Text composable overflows due to defined maxLines and it's made selectable, selecting the text causes a crash if the handle is moved outside text boundaries. The reason is that TextLayoutResult can still return the position of non-rendered characters when queried by the gesture handlers.

Instead of modifying selection behavior, this change limits where the selection handles can be drawn by adding a calculated `lastVisibleOffset` value. Selection handles are always coerced between 0 and `lastVisibleOffset` to make sure that handles are never tried to be drawn where invisible (overflown) characters were supposed to be placed.

Fix: b/262428141
Test: ./gradlew :compose:foundation:foundation:cAT
Change-Id: I273e3673ca00079953ec5aeda52aa009d8e73e25


GitOrigin-RevId: 1b849a6d08abfdf0adc73ebedbfa2d4932d01e5e