Make Marquee animate even when animations are disabled.
Animations can be disabled in the system a11y settings. However, marquee
is a load-bearing animation that typically occupies a very small space
in the screen, and should animate even when animations are disabled. See
discussion on bug for more information.

Fixes: b/262298306
Bug: b/265177763
Test: manually
Test: Not possible to test yet, see b/265177763. But ignored test added
    in BasicMarqueeTest.
Relnote: "`Modifier.basicMarquee` now animates even when animations are
    disabled in the system settings."
Change-Id: I23389686f06f2bf749a5f0326d0a33e6da23a1ec


GitOrigin-RevId: 4f02a054cdd109bb10537e0705ad09badff7945a