Calling requestLayout() when the ComposeView size might change
If LayoutNode can potentially resize the root we should call requestLayout() so our parent View can react on this change on the same frame. If instead we just call invalidate() and remeasure inside dispatchDraw() this will cause inconsistency as the Compose content will already have the new size, but the View hierarchy will react only on the next frame.

Test: ResizingComposeViewTest
Fixes: 264292272
Fixes: 173806298
Change-Id: I2fe6f2d31f30a1dd79e44e16ac82252564e63779


GitOrigin-RevId: 3188f33a19ba55f04004b769c6040eb8dd04eda0