Add reusable AndroidView API
This CL adds an overload AndroidView that allows View objects to be
reused in different parts of the composition. The new API leverages
`ReusableComposeNode` to opt-into automatic recycling from the composer.

The new API offers hooks to allow users to prepare their Views for
reuse, and also gives more information about when the View is being
released by Compose.

Relnote: """Added an overload of `AndroidView` composable function,
    which allows View instances to be reused when their node in the
    composition is discarded and reused in a compatible way. This is
    especially useful for LazyRows and LazyColumns of Views.
    """

Bug: 230099236
Test: ./gradlew compose:ui:ui:cAT
Test: ScrollingAndroidViewsDemo (manual)
Change-Id: I3f10db5de1b7699964274e0d25f4aad324865dca


GitOrigin-RevId: 9380a29f6c6ec2cf4b239997927eb3ff8b36342f