Add ScrollState.viewportSize in order to support scrollbars.
For a scrollbar to correctly size its "thumb" it needs to know the ratio between the size of the visible part of the scrolled content (viewport) and the size of the entire content. Additionally, to be able to scroll a "page" (when clicked on the track) it needs the size of the viewport. Unfortunately this information is not currently available in ScrollState.

In our current, incorrect, implementation, the scrollbar assumes that its own size (e.g. height for a vertical scrollbar) is also the size of the viewport. This is correct most of the time, so it has went unnoticed until now.

We considered several options to add this information:
1. Simply add a `viewportSize` property to ScrollState to indicate the visible size of the content. The size of the entire content can be calculated with `maxValue + viewportSize` (the case where the content is smaller than the viewport is irrelevant for a scrollbar).
2. Analogously to LazyListState add a `layoutInfo` property to ScrollState and put `viewportSize` in it. Possibly add `contentSize` there too. The latter would create the implicit dependency that `maxValue == max(0, contentSize - viewportSize)`.

The first option is currently our favorite, due to its simplicity, and because we don't foresee more fields being added into the `ScrollableLayoutInfo` suggested in option 2.

Test: unit tests

Change-Id: I16e79d7b8a286fda98b535efcb457e66ade60751


GitOrigin-RevId: 59fe7eadc17c215e09d2232c0c6e07705119f982