Add script to generate packages with Composable functions
Create a script (generate_compose_packages.py) to generate a text file
containing all the framework packages containing at least one composable
function. The script can also check if the packages stored in the
compose_packages_list.txt file match the current Compose packages.

When updating the packages list, the script will also regenerate the
PackageHashes.kt file, which calculates the hash for each package and
exposes them through a list in Kotlin (systemPackages).

The idea is to make this script run in presubmit, to make sure that we
always keep the list of packages up-to-date.

Bug: 258665704
Test: N/A. ui-inspector tests are expected to pass.

Change-Id: I2a86b107bad389ba20fc441986d35dab6099095d


GitOrigin-RevId: c62c4bbd4ddac2183a95bee2f713319fa21d9f2d