Add `@TestOnly` annotation to `Composer.disableSourceInformation()`
`disableSourceInformation()` can only be safely called in a test and
should never be called outside a test as it will change the result
of the composition in unpredictable ways. For this reason it should
have been marked `@TestOnly`.

Also added documentation about when this function can be called and
what the function does.

Test: N/A - annotation and documentation change only
Fixes: 260595475
Relnote: "Added `@TestOnly` to `Composer.disableSourceInformation()`
as this function is only safe to call in a test."

Change-Id: I896c84a91e1ec01f0107c52def0900196e2cde45


GitOrigin-RevId: fec2853902205da52649204078b560e2b1e1df5e