Fix issues with custom focus exit
Fix a bug in nested focusGroups where a custom exit property
with FocusRequester.Cancel allows focus to leave the focus group.

This issue was hidden by the fact that we were not distinguishing between a focus search that didnt
find a child and a focus search that encountered FocusRequester.Cancel. TwoDimensionalFocusSearch
would return false in both these cases, and a nested focusGroup would perceive a "Cancel" as "no focusable found in subtree", and continue searching.

Also refactored the code so that FocusReuester.Cancel and FocusRequester.Default don't have default requestFocus()
implementations. This way whenever we fetch focus properties
we have to explicitly check if the value is
FocusRequester.Default and can't rely on default implementations
of FocusRequester.Cancel and FocusRequester.Default which
caused the bug being addressed here.

Bug: 267827550
Test: ./gradlew compose:ui:ui:cC -P android.testInstrumentationRunnerArguments.package=androidx.compose.ui.focus.TwoDimensionalFocusTraversalImplicitExitTest
Change-Id: I6e085a536a4a829cb72f09f4dacf7fb71c66985c


GitOrigin-RevId: b33ca1f918ca074297e9320ec504a378e108eee6