Fix flaky grid prefetch tests
In grid tests we were waiting for the first cell of the row to be prefetched before running the assertion for the whole row. However in reality it is possible that the first cell will be prefetched on one frame, and the second cell will only be composed in the next frames. So we have to wait for both cells to be fully prefetched.

Fixes: 268210994
Test: run those tests
Change-Id: Iaa1c96577c6fa73d36b7c42cc9e4292b5c1f6526


GitOrigin-RevId: 69f17cdfd9792d432299a865a8400f20feb177d2