Use initialized value from LiveData for first composition
Previously, observeAsState(initial) would always show
the initial value on the first composition, even if the
LiveData already had a valid value set on it.

By using the LiveData.isInitialized check, we can
distinguish between the cases where no value has ever
been set on the LiveData (which case we'll use the
initial value passed to the function) versus the cases
where a value has already been set on the LiveData.

This ensures that even on the first composition, a
LiveData with a valid value will have the correct
State emitted.

Relnote: "The `LiveData.observeAsState(initial)` API
now correctly uses the `value` already set on the LiveData
in the first composition. The `initial` value passed to this
function is now only used if `LiveData.isInitialized`
returns false (i.e., no value has ever been set on the
`LiveData`). Note that for LiveData types that have a
non-null generic type (e.g., `LiveData<String>`), it is
your responsibility to ensure only non-null values are
set on the `LiveData`."
Test: new preferInitializedValueInFirstComposition test
BUG: 180450267
BUG: 214434197

Change-Id: Ie8f494ffcaed65c7ae8f08dabb8304be6721cd0f


GitOrigin-RevId: 3680e25632ab3511e476031f1f19df1fcfb08346