Introduce tryMutate method in InternalMutatorMutex
We are keeping a copy of Foundation's MutatorMutex in Material now to add an internal API for a bugfix.

tryMutate adds an opportunistic mutation API to MutatorMutex that allows to synchronously
attempt to mutate. This is useful for scenarios where immediate mutation is needed, but
coroutine dispatching typically happens in the next frame.

Test: Copied over MutatorMutexTest into InternalMutatorMutexTest, added
tryMutateBlockingSuspendsSubsequentMutate, tryMutateDoesNotOverrideActiveCaller
and tryMutateBlockingTryMutateLocks
Relnote: N/A

Change-Id: Ia7f6740e91928c0c9ba18edf6712766f334c092a


GitOrigin-RevId: ce6f12dee4313da43367325a0bde8374f86d9e66