Use JVM reflection for modifier inspector defaults
This reverts a change from 4e521e68699168fd1a7b5102a166733dc53ea07d
where we used kotlin-reflect to provide an automatic population of
Modifier.Node default inspectable properties. This now uses an
expect/actual declaration, which should make the default implementation
more available in general, and indirectly fixes an issue in Play on G3
where the dependency on kotlin-reflect was falsely detected as a
transitive dependency and prevented shipping.

Fixes: b/269496185
Test: ./gradlew :compose:ui:ui:tDUT
Change-Id: I4cbd7d08f1ccabe5e50917a7280d3450ed6c89df


GitOrigin-RevId: 24413c0426c208a71973f5514e9cbf8a5ed675a3