Make SetText action update buffer when in session & use SemanticsActions for test functions.
This avoids the internal buffer getting out of sync if both the SetText
action and IME commands are issued in the same frame, and allows us to
use semantics for the test functions (b/269633168).

This change also changes the performTextClearance,
performTextReplacement, and performTextSelection test functions to use
the respective semantics actions instead of sending EditCommands
directly to the session. This is more consistent with how most of the
test helper functions work, and makes it easier to decouple text from
the UI module (aosp/2406439).

Fixes: b/269633168
Fixes: b/269624358
Test: TextFieldTest
Test: All existing text field tests cover this.
Relnote: "`BasicTextField`'s `SetText` semantics action will now update
    the text buffer using the same code path as IME updates and the
    testing functions (e.g. `performTextReplacement`)."
Relnote: "Text testing functions `performTextClearance`,
    `performTextReplacement`, and `performTextSelection` now use
    `SemanticsActions`."
Change-Id: I0807d73975224ac5bf02fc232c1ab615f76c1c7d


GitOrigin-RevId: 9e006f2dd33d58bc3ead0613125de8d6953f2d67