Handle tab and enter key events in CoreTextField
Before this, the tab and enter keys are treated as typed events (like a letter)
and simply adds a tab or newline character to the text. Forwarding it to the
KeyMapping allows it to be treated specially instead when single-line,
and as appending a character in multi-line.

Additionally, adds a demo for text field focus changes.

Test: ./gradlew :compose:foundation:foundation:connectedAndroidTest -Pandroid.testInstrumentationRunnerArguments.class=androidx.compose.foundation.textfield.TextFieldFocusTest
Test: ./gradlew :compose:foundation:foundation:connectedAndroidTest -Pandroid.testInstrumentationRunnerArguments.class=androidx.compose.foundation.textfield.HardwareKeyboardTest
Fixes: b/197978180
Fixes: b/209930549
Fixes: b/199919712
Fixes: b/200596944
Fixes: b/202243080
Fixes: b/237676253
Change-Id: I53ad695f0aaf477291ea255663bd7e0ef498dcfe


GitOrigin-RevId: c806153a6050623d5c390642b75cdb2e4b4c81e8