Track AndroidView's wrapped View in a non-null Val
AndroidViewHolder's `view` property never actually changes to a
different View instance. Now that we're being more aggressive about
detaching View instances, it's leading to unexpected NPEs.

This CL should address that in a slightly roundabout way. We can keep
the View property in a non-null `val` to make it more representative of
reality. Where we're currently assigning the View to null, we now just
need to detach and reattach the View.

Bug: b/271156218
Test: ./gradlew compose:ui:ui:cAT
Change-Id: I04c7e77554b30234fba929b20c0c18e4be0b13d0


GitOrigin-RevId: 9a842a7ab4ec6aaea0c50c016a1c6ed5c0710ee4