Avoid !! in AndroidView onInitializeAccessibilityNodeInfo
The delegate in question was observed to crash with NPE in b/2606620990,
in a scenario involving ExoPlayer.  This code can be rewritten in a
fully null-safe way by removing the dependency on SemanticsNode.

New test androidViewAccessibilityDelegate adds coverage for the
onInitializeAccessibilityNodeInfo block, but isn't strictly speaking
testing the change, as I wasn't able to find any crashing scenario.  The
crash either was fixed already in 1.4 due to changes on the ModifierNode
side, or it's a very specific scenario that I didn't imagine.

Test: AndroidViewTest.androidViewAccessibilityDelegate
Bug: 260620990
Change-Id: I62ef88b4eb412a48daaf4bf5c998e2f5814591a9


GitOrigin-RevId: 16f722dbd21c8d9a2009d4738d21d3835429bc4c