Small optimizations to PathParser
On a Pixel 6, parsing of paris_30k.xml (17 MiB) goes from 1.32s
down to 1.29s. The optimization flips a bit in the int representation
of characters to test if a character is either a lower case letter
or upper case letter using a single condition. The generated Kotlin
bytecode looks a lot better than the source suggests (constants like
'e'.code are completely inlined using an integer constant for
instance).

Test: VectorTest
Test: PathParserTest
Test: ImageVectorTest
Change-Id: I6fdf3ba527c7784bfc3dd40bf58a925ca9736e2d


GitOrigin-RevId: 7596bf04fc9e0814af8d961aa6377566bcf976c9