Remove the `fill` in `SlotTable.moveSlotGapTo()` and move it to close
The `fill` `moveSlotGapTo` keeps the gap in the slot array to
always contains nulls. However this constraint is only necessary
to keep when the slot table is not being written to. This change
moves the clearing to the slot table to `close()`.

See the original change request aosp/2496017 for more details.

This change does not directly add tests but changes
SlotTable.verifyWellFormed to verify the constraing that the
slots gap only contains nulls. verifyWellFormed is called
SlotTableTests and CompositionTests.

Fixes: 274136013
Test: ./gradlew :compose:r:r:tDUT

Change-Id: Iccc23740ba06a0aa952a5f6f52999de247ce62c5


GitOrigin-RevId: 81f9935403b23876fb0ca4870606462b2e3a34d4