Simplify the BasicTextField2 package structure.
Everything for the new text fields should now live in text2/input/,
except for BasicTextField2 itself. BTF2 needs to be in text2/ because
the current BasicTextField is in text/, so that's where the new one
should live after the rewrite is finished. All other public APIs are in
text2/input/, and all internal/implementation code is in
text2/input/internal/.

Test: n/a, just moving files around
Relnote: n/a
Change-Id: I0164a2b4baed62a30b413e0309895600e939fb63


GitOrigin-RevId: 715f1edc33e56dd6b381a9bf177d67020bb2750a