Fix reused modifier nodes not being re-attached
The current implementation of LayoutNode requires Modifiers to be
reassigned after the node is detached and reattached. Otherwise, the
modifiers will remain in an invalid state where they are never
re-attached even though their parent LayoutNode is active.

This right now seems to be only manifesting as a crash when using
the reusable AndroidView overload, since AndroidView does not always
reassign its modifiers after being reused. This CL causes modifiers to
always be reactivated when the LayoutNode is reused.

Test: Added test to ModifierNodeReuseAndDeactivationTest
Bug: b/275919849
Change-Id: Ib67ee784e908b4aea5c1de061e448267b97d2349


GitOrigin-RevId: 62e7039dbf7f5182fe6ecbf7b4ffb0a22cdff03a