[M2] Remove swipeAnchors API
Updating the anchors using SwipeableV2State#updateAnchors is safe at any time, although recommended to be called during layout for layout-dependent anchors.
This means we can recommend using onSizeChanged instead of the custom modifier and simplify the API surface.

This also updates the existing tests to remove the test helpers.

Test: Updated existing tests
Relnote: N/A
Bug: 277049096
Change-Id: Ib7c5048dbbcf70e9f130e43741e5358f96bfc996


GitOrigin-RevId: 571d161cd99c706ed4d9dd6c8c8fbacb42deb23f