Don't capture stack trace when cancelling Modifier.Node coroutine scope.
Capturing stack traces is expensive, and modifier node cancellation
happens at a well-defined time, so there's little value in having a
stack trace for it. We unconditionally avoid capturing a stack trace in
that case in a similar way to how the coroutines library does it when
cancel is called without a message.

See https://chat.google.com/room/AAAAbSAVMlo/Pm3MPxoL2dc

Test: ModifierNodeCoroutineScopeTest
Relnote: n/a
Change-Id: I4c3d495482013530320235f1f6305ea68bbdcde7


GitOrigin-RevId: db2ade187851d6651a5670230b3ab8c38645f994