Make Modifier.Node.detach() not pass a message at all when cancelling coroutine scope.
This is a different way to make sure that detach() doesn't capture a
stack trace when cancelling the scope. The initial fix introduced a new
exception type in aosp/2535088, but after looking at the rest of the
androidx codebase, we almost never cancel a scope with a message so this
is more idiomatic.

See https://chat.google.com/room/AAAAbSAVMlo/Pm3MPxoL2dc

Test: ModifierNodeCoroutineScopeTest
Relnote: n/a
Change-Id: I965b6bd1ee2c2bf1e6b5fb0d344be81df72dfc49


GitOrigin-RevId: e88e8fa3c0a29a2cfa4992832f8ab2ad0db590cb