Add lint check to specialize mutableStateOf calls
This CL adds an inspection to warn users when they are creating a state
using the generic `mutableStateOf<T>()` function that could be
specialized to a different state builder function which will not box
type `T`. To make this conversion, a few things must be true (which
this inspection looks for when reporting errors):
 - `T` must not be nullable
 - `T` must have a known specialized state builder function
 - The snapshot mutation policy argument must not be specified or must
   be the `structuralEqualityPolicy()`

For AndroidX, this inspection will be treated as an error.

Test: AutoboxingStateCreationDetectorTest
Change-Id: I4781c5b7bb48d177619fe5dece3b23f3989d41ee


GitOrigin-RevId: 73507b7fb76029a9e341b8bc7e12f27f7619a55e