Remove allocations and copies from vector APIs.
PathComponent and GroupComponent relied on PathParser to convert a
List<PathNode> into a Path. PathParser required many unncessary
allocations (normally used for actual parsing of string representations
of paths), so this change extracts the List<PathNode> to Path conversion
code into a top-level extension function on List<PathNode>.

Additionally, toPath() is rewritten as a single function that performs
all the conversions within its internal switch table instead of
relying on extension functions on PathNode subclasses. This allows the
removal of other allocations since the function can now be used without
being part of a class to hold said allocations.

An added benefit is a simplification of the call sites as they don't
need to manage a PathParser instance anymore.

Test: VectorTest
Test: compose.integration-tests.demo
Test: PathParserTest
RelNote: Improved performance and reduce allocations in Vector APIs

Change-Id: I906cb6e7d7a950b11c27f414385aa7f9c9f514aa


GitOrigin-RevId: 4e166852b6349699d2a327ef295d97ab9200a814